% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Arre.aura-data.R
\docType{data}
\name{Arre.aura}
\alias{Arre.aura}
\title{Acoustic recording of \emph{Arremon aurantiirostris}.}
\format{One .wav file: \describe{

\item{Arre.aura}{\emph{Arremon aurantiirostris} recording}

 }}
\source{
\url{http://www.xeno-canto.org}
}
\usage{
data(Arre.aura)
}
\description{
Contained within "recordings" data set, with \emph{Phaeocroa cuvierii}
and \emph{Phaethornis longirostris} recordings.
}
\keyword{datasets}

