% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autodetec.R
\name{auto_detec}
\alias{auto_detec}
\title{alternative name for \code{\link{autodetec}}}
\usage{
auto_detec(
  X = NULL,
  wl = 512,
  threshold = 15,
  parallel = 1,
  power = 1,
  output = "data.frame",
  thinning = 1/10,
  path = NULL,
  pb = TRUE,
  ssmooth = NULL,
  bp = NULL,
  flist = NULL,
  hold.time,
  envt = NULL,
  msmooth = NULL,
  osci = NULL,
  xl = NULL,
  picsize = NULL,
  res = NULL,
  flim = NULL,
  ls = NULL,
  sxrow = NULL,
  rows = NULL,
  mindur = NULL,
  maxdur = NULL,
  redo = NULL,
  img = NULL,
  it = NULL,
  set = NULL,
  smadj = NULL,
  pal = NULL,
  fast.spec = NULL
)
}
\description{
alternative name for \code{\link{autodetec}}
}
\details{
see \code{\link{autodetec}} for documentation. \code{\link{autodetec}} will be deprecated in future versions.
}
\keyword{internal}
