% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checksels.R
\name{check_sels}
\alias{check_sels}
\title{alternative name for \code{\link{checksels}}}
\usage{
check_sels(X = NULL, parallel = 1, path = NULL, check.header = FALSE,
  pb = TRUE, wav.size = FALSE)
}
\description{
alternative name for \code{\link{checksels}}
}
\details{
see \code{\link{checksels}} for documentation. \code{\link{checksels}} will be deprecated in future versions.
}
\keyword{internal}
