% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav_check.R
\name{check_wavs}
\alias{check_wavs}
\title{alternative name for \code{\link{wav_check}}}
\usage{
check_wavs(X = NULL, path = NULL)
}
\description{
alternative name for \code{\link{wav_check}}
}
\details{
see \code{\link{wav_check}} for documentation. \code{\link{check_wavs}} will be deprecated in future versions.
}
\keyword{internal}
