% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sels.R
\name{checksels}
\alias{checksels}
\title{alternative name for \code{\link{check_sels}}}
\usage{
checksels(
  X = NULL,
  parallel = 1,
  path = NULL,
  check.header = FALSE,
  pb = TRUE,
  wav.size = FALSE,
  verbose = TRUE
)
}
\description{
alternative name for \code{\link{check_sels}}
}
\details{
see \code{\link{check_sels}} for documentation. \code{\link{checksels}} will be deprecated in future versions.
}
\keyword{internal}
