% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_wavs.R
\name{checkwavs}
\alias{checkwavs}
\title{alternative name for \code{\link{check_wavs}}}
\usage{
checkwavs(X = NULL, path = NULL)
}
\description{
alternative name for \code{\link{check_wavs}}
}
\details{
see \code{\link{check_wavs}} for documentation. \code{\link{checkwavs}} will be deprecated in future versions.
}
\keyword{internal}
