% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.spectro.R
\name{color_spectro}
\alias{color_spectro}
\title{alternative name for \code{\link{color.spectro}}}
\usage{
color_spectro(
  wave,
  wl = 512,
  wn = "hanning",
  ovlp = 70,
  dB = "max0",
  collevels = NULL,
  selec.col = "red2",
  col.clm = NULL,
  base.col = "black",
  bg.col = "white",
  cexlab = 1,
  cexaxis = 1,
  tlab = "Time (s)",
  flab = "Frequency (kHz)",
  title = NULL,
  axisX = TRUE,
  axisY = TRUE,
  flim = NULL,
  rm.zero = FALSE,
  X = NULL,
  fast.spec = FALSE,
  t.mar = NULL,
  f.mar = NULL,
  interactive = NULL,
  add = FALSE
)
}
\description{
alternative name for \code{\link{color.spectro}}
}
\details{
see \code{\link{color.spectro}} for documentation. \code{\link{color.spectro}} will be deprecated in future versions.
}
\keyword{internal}
