% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods.R
\name{compare.methods}
\alias{compare.methods}
\title{alternative name for \code{\link{compare_methods}}}
\usage{
compare.methods(
  X = NULL,
  flim = NULL,
  bp = NULL,
  mar = 0.1,
  wl = 512,
  ovlp = 90,
  res = 150,
  n = 10,
  length.out = 30,
  methods = NULL,
  it = "jpeg",
  parallel = 1,
  path = NULL,
  sp = NULL,
  custom1 = NULL,
  custom2 = NULL,
  pb = TRUE,
  grid = TRUE,
  clip.edges = TRUE,
  threshold = 15,
  na.rm = FALSE,
  scale = FALSE,
  pal = reverse.gray.colors.2,
  img = TRUE,
  ...
)
}
\description{
alternative name for \code{\link{compare_methods}}
}
\details{
see \code{\link{compare_methods}} for documentation. \code{\link{compare.methods}} will be deprecated in future versions.
}
\keyword{internal}
