% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coor.sing-data.R
\docType{data}
\name{coor.sing}
\alias{coor.sing}
\title{Selections of coordinated singing events from lekking \emph{Phaethornis longirostris} (Long-billed Hermit hummingbird).}
\format{\describe{

  \item{coor.sing}{Selections of coordinated singing}

  }}
\usage{
data(coor.sing)
}
\description{
Selections of coordinated singing events from lekking \emph{Phaethornis longirostris} (Long-billed Hermit hummingbird).
}
\keyword{datasets}

