% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor_test.R
\name{coor.test}
\alias{coor.test}
\title{alternative name for \code{\link{coor_test}}}
\usage{
coor.test(
  X = NULL,
  iterations = 1000,
  ovlp.method = "count",
  randomization = "keep.gaps",
  less.than.chance = TRUE,
  parallel = 1,
  pb = TRUE,
  rm.incomp = FALSE,
  cutoff = 2,
  rm.solo = FALSE
)
}
\description{
alternative name for \code{\link{coor_test}}
}
\details{
see \code{\link{coor_test}} for documentation. \code{\link{coor.test}} will be deprecated in future versions.
}
\keyword{internal}
