% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor.test.R
\name{coor_test}
\alias{coor_test}
\title{alternative name for \code{\link{coor.test}}}
\usage{
coor_test(X = NULL, iterations = 1000, less.than.chance = TRUE,
  parallel = 1, pb = TRUE, rm.imcomp = FALSE, cutoff = 2,
  rm.solo = FALSE)
}
\description{
alternative name for \code{\link{coor.test}}
}
\details{
see \code{\link{coor.test}} for documentation. \code{\link{coor.test}} will be deprecated in future versions.
}
\keyword{internal}
