% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cryp.soui-data.R
\docType{data}
\name{Cryp.soui}
\alias{Cryp.soui}
\title{Acoustic recording of \emph{Crypturellus soui} (Little Tinamou).}
\format{One .wav file: \describe{
  
  \item{Cryp.soui}{\emph{Crypturellus soui} recording}
  
  }}
\source{
\url{http://www.xeno-canto.org/154190}
}
\usage{
data(Cryp.soui)
}
\description{
Acoustic recording of \emph{Crypturellus soui} (Little Tinamou).
}
\keyword{datasets}
