% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_cut.R
\name{cut_sels}
\alias{cut_sels}
\title{alternative name for \code{\link{sel_cut}}}
\usage{
cut_sels(
  X,
  mar = 0.05,
  parallel = 1,
  path = NULL,
  dest.path = NULL,
  pb = TRUE,
  labels = c("sound.files", "selec"),
  overwrite = FALSE,
  norm = FALSE,
  ...
)
}
\description{
alternative name for \code{\link{sel_cut}}
}
\details{
see \code{\link{sel_cut}} for documentation. \code{\link{cut_sels}} will be deprecated in future versions.
}
\keyword{internal}
