% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_DTW.R
\name{dfDTW}
\alias{dfDTW}
\title{alternative name for \code{\link{freq_DTW}}}
\usage{
dfDTW(
  X = NULL,
  type = "dominant",
  wl = 512,
  wl.freq = 512,
  length.out = 20,
  wn = "hanning",
  ovlp = 70,
  bp = c(0, 22),
  threshold = 15,
  threshold.time = NULL,
  threshold.freq = NULL,
  img = TRUE,
  parallel = 1,
  path = NULL,
  ts.df = NULL,
  img.suffix = "dfDTW",
  pb = TRUE,
  clip.edges = TRUE,
  window.type = "none",
  open.end = FALSE,
  scale = FALSE,
  frange.detec = FALSE,
  fsmooth = 0.1,
  adjust.wl = TRUE,
  ...
)
}
\description{
alternative name for \code{\link{freq_DTW}}
}
\details{
see \code{\link{freq_DTW}} for documentation. \code{\link{dfDTW}} will be deprecated in future versions.
}
\keyword{internal}
