% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_performace.R
\name{diagnose_detection}
\alias{diagnose_detection}
\title{Describes the performance of a signal detection procedure referenced in a selection table}
\usage{
diagnose_detection(reference, detection, by.sound.file = FALSE, 
time.diagnostics = FALSE)
}
\arguments{
\item{reference}{Data frame or 'selection.table' with the reference selections (start and end of the signals) that will be used to evaluate the performance of the detection, represented by those selections in 'detection'. Must contained at least the following columns: "sound.files", "selec", "start" and "end".}

\item{detection}{Data frame or 'selection.table' with the detections (start and end of the signals) that will be compared against the 'reference' selections. Must contained at least the following columns: "sound.files", "selec", "start" and "end".}

\item{by.sound.file}{Logical argument to control whether performance diagnostics are summarized across sound files (when \code{by.sound.file = FALSE}, when more than 1 sound file is included in 'reference') or shown separated by sound file. Default is \code{FALSE}.}

\item{time.diagnostics}{Logical argument to control if diagnostics related to the duration of the signals ("mean.duration.true.positives", "mean.duration.false.positives", "mean.duration.false.negatives" and "proportional.duration.true.positives") are returned (if \code{TRUE}). Default is \code{FALSE}.}
}
\value{
A data frame including the following detection performance diagnostics:
\itemize{
 \item \code{true.positives}: number of detections that correspond to signals referenced in 'reference'. Matching is defined as some degree of overlap in time. In a perfect detection routine it should be equal to the number of rows in 'reference'. 
 \item \code{false.positives}: number of detections that don't match any of the signals referenced in 'reference'. In a perfect detection routine it should be 0.    
 \item \code{false.negatives}: number of signals in 'reference' that were not detected (not found in 'detection'. In a perfect detection routine it should be 0. 
 \item \code{split.positives}: number of signals referenced in 'reference' that were overlapped by more than 1 detection (i.e. detections that were split). In a perfect detection routine it should be 0.  
 \item \code{mean.duration.true.positives}: mean duration of true positives (in s). Only included when \code{time.diagnostics = TRUE}.  
 \item \code{mean.duration.false.positives}: mean duration of false positives (in s). Only included when \code{time.diagnostics = TRUE}.
 \item \code{mean.duration.false.negatives}: mean duration of false negatives (in s). Only included when \code{time.diagnostics = TRUE}.
 \item \code{proportional.duration.true.positives}: ratio of total duration of true positives to the total duration of signals referenced in 'reference'. In a perfect detection routine it should be 1. Only included when \code{time.diagnostics = TRUE}.
 \item \code{sensitivity}: Proportion of signals referenced in 'reference' that were detected. In a perfect detection routine it should be 1.
 \item \code{specificity}: Proportion of detections that correspond to signals referenced in 'reference' that were detected. In a perfect detection routine it should be 1.
 }
}
\description{
\code{diagnose_detection} describes the performance of a signal detection procedure referenced in a selection table
}
\details{
The function describes the performance of a signal detection procedure referenced in a selection table by comparing it to a reference selection table in which all signals of interest have been selected.
}
\examples{
{
# load data
data(list = c("Phae.long1", "Phae.long2", "lbh_selec_reference"))

# save sound files
writeWave(Phae.long4, file.path(tempdir(), "Phae.long4.wav")) 
writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"))

# run detection with auto_detec
ad1 <- auto_detec(path = tempdir(), flist = unique(lbh_selec_reference$sound_files),  
ssmooth = 1200, pb = FALSE, threshold =22, mindur = 0.10, maxdur = 0.18, 
bp = c(3, 9), power = 1)

# summarizing across sound files
diagnose_detection(reference = lbh_selec_reference, detection = ad1, 
by.sound.file = FALSE)

# by sound file
diagnose_detection(reference = lbh_selec_reference, detection = ad1, 
by.sound.file = TRUE)

#' # by sound file including time diagnostics
diagnose_detection(reference = lbh_selec_reference, detection = ad1, 
by.sound.file = TRUE, time.diagnostics = TRUE)
}
}
\references{
{
Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.
}
}
\seealso{
\code{\link{auto_detec}}, \code{\link[monitoR]{findPeaks}}
}
\author{
Marcelo Araya-Salas \email{marcelo.araya@ucr.ac.cr})
}
