% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.raven.R
\name{exp_raven}
\alias{exp_raven}
\title{Export raven selections}
\usage{
exp_raven(X, path = NULL, file.name = NULL, khz.to.hz = TRUE)
}
\arguments{
\item{X}{Data frame with containing columns for sound file (sound.files), selection (selec), start and end time of signals ('start' and 'end') and low and high frequency ('low.freq' and 'high.freq', optional). See example data 'selec.table'.}

\item{path}{A character string indicating the path of the directory in which to look for the text files. 
If not provided (default) the function saves the file into the current working directory. Default is \code{NULL}).}

\item{file.name}{Name of the output .txt file.}

\item{khz.to.hz}{Logical. Controls if frequency variables should be converted from kHz (the unit used by warbleR) to Hz (the unit used by Raven). Default if \code{TRUE}.}
}
\value{
A single data frame with information of the selection files. If all.data argument is set to \code{FALSE}) the data 
frame contains the following columns: selec, start, end, and selec.file. If sound.file.col is provided the data frame
will also contain a 'sound.files' column. In addition, all rows with duplicated data are removed. This is useful when 
both spectrogram and waveform views are included in the Raven selection files. If all.data is set to \code{TRUE} then all 
columns in selection files are returned.
}
\description{
\code{exp_raven} exports selection tables as Raven selection data in .txt format.
}
\details{
The function exports selection tables used by warbleR to raven selection data in .txt format.
}
\examples{
\dontrun{
# First set temporary folder
setwd(tempdir())

# Load data
data("selec.table")

# Select data for a single sound file
st1 <- selec.table[selec.table$sound.files == "Phae.long1.wav",]

# Export data
exp_raven(st1, file.name = "Phaethornis warbleR examples")
}

}
\author{
Marcelo Araya-Salas (\email{araya-salas@cornell.edu})
}
\seealso{
\code{\link{imp.raven}}; \code{\link{imp.syrinx}}
}

