% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffDTW.R
\name{ff_DTW}
\alias{ff_DTW}
\title{alternative name for \code{\link{ffDTW}}}
\usage{
ff_DTW(X, wl = 512, length.out = 20, wn = "hanning", ovlp = 70,
  bp = c(0, 22), threshold = 5, img = TRUE, parallel = 1, path = NULL,
  img.suffix = "ffDTW", pb = TRUE, clip.edges = TRUE,
  window.type = "none", open.end = FALSE, scale = FALSE, ...)
}
\description{
alternative name for \code{\link{ffDTW}}
}
\details{
see \code{\link{ffDTW}} for documentation. \code{\link{ffDTW}} will be deprecated in future versions.
}
\keyword{internal}
