% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtersels.R
\name{filter_sels}
\alias{filter_sels}
\title{alternative name for \code{\link{filtersels}}}
\usage{
filter_sels(
  X,
  path = NULL,
  lspec = FALSE,
  img.suffix = NULL,
  it = "jpeg",
  incl.wav = TRUE,
  missing = FALSE,
  index = FALSE
)
}
\description{
alternative name for \code{\link{filtersels}}
}
\details{
see \code{\link{filtersels}} for documentation. \code{\link{filtersels}} will be deprecated in future versions.
}
\keyword{internal}
