% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sels.R
\name{filtersels}
\alias{filtersels}
\title{alternative name for \code{\link{filter_sels}}}
\usage{
filtersels(
  X,
  path = NULL,
  lspec = FALSE,
  img.suffix = NULL,
  it = "jpeg",
  incl.wav = TRUE,
  missing = FALSE,
  index = FALSE
)
}
\description{
alternative name for \code{\link{filter_sels}}
}
\details{
see \code{\link{filter_sels}} for documentation. \code{\link{filtersels}} will be deprecated in future versions.
}
\keyword{internal}
