% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav_fix.R
\name{fix_wavs}
\alias{fix_wavs}
\title{alternative name for \code{\link{wav_fix}}}
\usage{
fix_wavs(
  checksels = NULL,
  files = NULL,
  samp.rate = NULL,
  bit.depth = NULL,
  path = NULL,
  mono = FALSE
)
}
\description{
alternative name for \code{\link{wav_fix}}
}
\details{
see \code{\link{wav_fix}} for documentation. \code{\link{fix_wavs}} will be deprecated in future versions.
}
\keyword{internal}
