% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frange.detec.R
\name{freq_range_detec}
\alias{freq_range_detec}
\title{alternative name for \code{\link{frange.detec}}}
\usage{
freq_range_detec(wave, wl = 512, fsmooth = 0.1, threshold = 10,
  dB.threshold = NULL, wn = "hanning", flim = c(0, 22), bp = NULL,
  fast.spec = FALSE, ovlp = 50, pal = reverse.gray.colors.2,
  widths = c(2, 1), main = NULL, plot = TRUE, all.detec = FALSE)
}
\description{
alternative name for \code{\link{frange.detec}}
}
\details{
see \code{\link{frange.detec}} for documentation. \code{\link{frange.detec}} will be deprecated in future versions.
}
\keyword{internal}
