% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lspec2pdf.R
\name{full_spec2pdf}
\alias{full_spec2pdf}
\title{alternative name for \code{\link{lspec2pdf}}}
\usage{
full_spec2pdf(keep.img = TRUE, overwrite = FALSE, parallel = 1,
  path = NULL, pb = TRUE)
}
\description{
alternative name for \code{\link{lspec2pdf}}
}
\details{
see \code{\link{lspec2pdf}} for documentation
}
\keyword{internal}
