% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec_funs.R
\name{imp.raven}
\alias{imp.raven}
\title{deprecated functions and function names}
\usage{
imp.raven(...)
}
\description{
deprecated functions and function names
}
\details{
functions that have been deprecated or moved to other packages.
}
\keyword{internal}
