% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflections.R
\name{inflections}
\alias{inflections}
\title{Count number of inflections in a frequency contour}
\usage{
inflections(X = NULL, parallel = 1, pb = TRUE)
}
\arguments{
\item{X}{data frame with the columns for "sound.files" (sound file name), "selec" (unique identifier for each selection) and columns for each of the frequency values of the contours. No other columns should be included}

\item{parallel}{Numeric. Controls whether parallel computing is applied.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{pb}{Logical argument to control progress bar and messages. Default is \code{TRUE}.}
}
\value{
A data frame with 3 columns: "sound.files", "selec" and "infls" (number of inflections).
}
\description{
\code{inflections} counts the number of inflections in a frequency contour (or any time series)
}
\details{
The function counts the number of inflections in a frequency contour.
}
\examples{
{
# Set temporary working directory

# get warbleR sound file examples
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "selec.table"))
writeWave(Phae.long1,"Phae.long1.wav")
writeWave(Phae.long2,"Phae.long2.wav")
writeWave(Phae.long3,"Phae.long3.wav")
writeWave(Phae.long4,"Phae.long4.wav") 

# measure frequency contours
dom.freq.ts <- dfts(X = selec.table)

# get number of inflections
inflections(X = dom.freq.ts)
}

}
\seealso{
\code{\link{dfts}}, \code{\link{trackfreqs}},
}
\author{
Marcelo Araya-Salas (\email{araya-salas@cornell.edu})
}
