\name{warbleR-internals}
\alias{spectroW}
\alias{spectro2}
\alias{rectw}
\title{
\code{warbleR} Internal Functions
}
\description{
These functions are used by other functions in the warbleR package, but are not intended to be called by users. Both are modified versions of \code{\link[seewave]{spectro}}.
The functions are:
\code{spectroW}
\code{spectro2}

\code{rectw} is a modified version of \code{\link[graphics]{rect}} to add density (cross-hatching lines)
}

\keyword{internal}

\note{
  These functions are not to be called by the user.
}