% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbh_selec_table-data.R
\docType{data}
\name{lbh_selec_table}
\alias{lbh_selec_table}
\title{Alternative name for \code{lbh_selec_table}}
\format{An object of class \code{data.frame} with 11 rows and 9 columns.}
\usage{
lbh_selec_table
}
\description{
\code{lbh_selec_table} alternative name for \code{lbh_selec_table}. \code{selec.table} will be deprecated in 
future versions.
}
\details{
Simulated coordinated singing events.
}
\keyword{datasets}
