% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_spectrogram2pdf.R
\name{lspec2pdf}
\alias{lspec2pdf}
\title{alternative name for \code{\link{full_spectrogram2pdf}}}
\usage{
lspec2pdf(
  keep.img = TRUE,
  overwrite = FALSE,
  parallel = 1,
  path = NULL,
  pb = TRUE
)
}
\description{
alternative name for \code{\link{full_spectrogram2pdf}}
}
\details{
see \code{\link{full_spectrogram2pdf}} for documentation
}
\keyword{internal}
