% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_table.R
\name{make.selection.table}
\alias{make.selection.table}
\title{Old name for \code{\link{selection_table}}}
\usage{
make.selection.table(
  X,
  max.dur = 10,
  path = NULL,
  whole.recs = FALSE,
  extended = FALSE,
  confirm.extended = TRUE,
  mar = 0.1,
  by.song = NULL,
  pb = TRUE,
  parallel = 1,
  ...
)
}
\description{
Old name for \code{\link{selection_table}}
}
\details{
see \code{\link{selection_table}} for documentation
}
\keyword{internal}
