% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_imgs.R
\name{move.imgs}
\alias{move.imgs}
\title{alternative name for \code{\link{move_imgs}}}
\usage{
move.imgs(
  from = NULL,
  to = NULL,
  it = "all",
  cut = TRUE,
  overwrite = FALSE,
  create.folder = TRUE,
  folder.name = "image_files",
  parallel = 1,
  pb = TRUE
)
}
\description{
alternative name for \code{\link{move_imgs}}
}
\details{
see \code{\link{move_imgs}} for documentation. \code{\link{move.imgs}} will be deprecated in future versions.
}
\keyword{internal}
