% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_find_peaks.R
\name{optimize_find_peaks}
\alias{optimize_find_peaks}
\title{Optimize the detection of signals based on a-priori detections}
\usage{
optimize_find_peaks(xc.output, reference, cutoffs = NULL, parallel = 1,
 pb = FALSE, by.sound.file = FALSE, previous.output = NULL)
}
\arguments{
\item{xc.output}{An object of class 'xcorr.output' (generated by \code{\link{cross_correlation}}) in which to optimize detections. Must refer to the same sound files as in 'X'.}

\item{reference}{Data frame or 'selection.table' with the reference selections (start and end of the signals) that will be used to evaluate the performance of the detection, represented by those selections in 'detection'. Must contained at least the following columns: "sound.files", "selec", "start" and "end".}

\item{cutoffs}{Numeric vector with the cutoff values to be tested for optimization. Default is \code{NULL}.}

\item{parallel}{Numeric. Controls whether parallel computing is applied.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{pb}{Logical argument to control progress bar and messages. Default is \code{TRUE}.}

\item{by.sound.file}{Logical to control if diagnostics are calculated for each sound file independently (\code{TRUE}) or for all sound files combined (\code{FALSE}, default).}

\item{previous.output}{Data frame with the output of a previous run of this function. This will be used to include previous results in the new output and avoid recalculating detection performance for parameter combinations previously evaluated.}
}
\value{
A data frame in which each row shows the result of a detection job for each cutoff value, including the following diagnostic metrics:
\itemize{
 \item \code{true.positives}: number of detections that correspond to signals referenced in 'X'. Matching is defined as some degree of overlap in time. In a perfect detection routine it should be equal to the number of rows in 'X'. 
 \item \code{false.positives}: number of detections that don't match any of the signals referenced in 'X'. In a perfect detection routine it should be 0.    
 \item \code{split.positives}: number of signals referenced in 'X' that were overlapped by more than 1 detection (i.e. detections that were split). In a perfect detection routine it should be 0.
 \item \code{sensitivity}: Proportion of signals referenced in 'X' that were detected. In a perfect detection routine it should be 1.
 \item \code{specificity}: Proportion of detections that correspond to signals referenced in 'X' that were detected. In a perfect detection routine it should be 1.
 }
}
\description{
Optimize the detection of signals based on a-priori detections
}
\details{
This function takes a selections data frame or 'selection_table' ('X') and the output of a \code{\link{auto_detec}} routine ('xc.output') and estimates the detection performance for different detection parameter combinations. This is done by comparing the position in time of the detection to those of the reference selections in 'X'. The function returns several diagnostic metrics to allow user to determine which parameter values provide a detection that more closely matches the selections in 'X'. Those parameters can be later used for performing a more efficient detection using \code{\link{auto_detec}}.
}
\examples{
{
# Save sound files to temporary working directory
data(list = c("Phae.long4", "Phae.long2", "lbh_selec_reference"))
writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"))
writeWave(Phae.long4, file.path(tempdir(), "Phae.long4.wav"))

# select data for 2 recordings
X <- lbh_selec_reference[grep("2|4", lbh_selec_reference$sound.files), ]

# create matrix to determine selections to be used as templates for cross correlation detection
cmp.mt <- cbind(paste(X$sound.files, X$selec, sep = "-"), X$sound.files)

# run cross correlation
xc.output <- cross_correlation(X = X, output = "list", 
compare.matrix = cmp.mt, ovlp = 50, wl = 300, path = tempdir())

# test 6 cutoff values
optimize_find_peaks(reference = X, xc.output = xc.output, 
cutoffs = seq(0.1, 0.6, by = 0.1))
}

}
\references{
{
Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.
}
}
\seealso{
\code{\link{optimize_auto_detec}}, \code{\link{find_peaks}}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr}).
}
