% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_overlap.R
\name{ovlp_sels}
\alias{ovlp_sels}
\title{alternative name for \code{\link{sel_overlap}}}
\usage{
ovlp_sels(
  X,
  index = FALSE,
  pb = TRUE,
  max.ovlp = 0,
  relabel = FALSE,
  drop = FALSE,
  priority = NULL,
  priority.col = NULL,
  unique.labs = TRUE,
  indx.row = FALSE,
  parallel = 1
)
}
\description{
alternative name for \code{\link{sel_overlap}}
}
\details{
see \code{\link{sel_overlap}} for documentation. \code{\link{ovlp_sels}} will be deprecated in future versions.
}
\keyword{internal}
