% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Phae.cuvi-data.R
\docType{data}
\name{Phae.cuvi}
\alias{Phae.cuvi}
\title{Acoustic recording of \emph{Phaeocroa cuvierii}.}
\format{One .wav file: \describe{

\item{Phae.cuvi}{\emph{Phaeocroa cuvierii} recording}

 }}
\source{
\url{http://www.xeno-canto.org}
}
\usage{
data(Phae.cuvi)
}
\description{
Contained within "recordings" data set, with \emph{Arremon
aurantiirostris}, and \emph{Phaethornis longirostris} recordings.
}
\keyword{datasets}

