% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Phae.long1-data.R
\docType{data}
\name{Phae.long1}
\alias{Phae.long1}
\title{Acoustic recording #1 of \emph{Phaethornis longirostris}.}
\format{One .wav file: \describe{

  \item{Phae.long1}{\emph{Phaethornis longirostris} #1 recording}

  }}
\source{
\url{http://www.xeno-canto.org}
}
\usage{
data(Phae.long1)
}
\description{
Contained within "recordings" data set, with \emph{Arremon aurantiirostris},
\emph{Phaeocroa cuvierii} and other \emph{Phaethornis longirostris} recordings.
}
\keyword{datasets}

