% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Phae.long3-data.R
\docType{data}
\name{Phae.long3}
\alias{Phae.long3}
\title{Acoustic recording #3 of \emph{Phaethornis longirostris}.}
\format{One .wav file: \describe{

  \item{Phae.long3}{\emph{Phaethornis longirostris} #3 recording}

  }}
\source{
\url{http://www.xeno-canto.org}
}
\usage{
data(Phae.long3)
}
\description{
Contained within "recordings" data set, with \emph{Arremon aurantiirostris},
\emph{Phaeocroa cuvierii} and other \emph{Phaethornis longirostris} recordings.
}
\keyword{datasets}

