% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_detec.R
\name{print.autodetec.output}
\alias{print.autodetec.output}
\title{print method for class \code{autodetec.output}}
\usage{
\method{print}{autodetec.output}(x, ...)
}
\arguments{
\item{x}{Object of class \code{autodetec.output}, generated by \code{\link{auto_detec}}.}

\item{...}{further arguments passed to or from other methods. Ignored when printing selection tables.}
}
\description{
print method for class \code{autodetec.output}
}
\keyword{internal}
