% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks.R
\name{print.find_peaks.output}
\alias{print.find_peaks.output}
\title{print method for class \code{xcorr.output}}
\usage{
\method{print}{find_peaks.output}(x, ...)
}
\arguments{
\item{x}{Object of class \code{find_peaks.output}, generated by \code{\link{find_peaks}}.}

\item{...}{further arguments passed to or from other methods. Ignored when printing selection tables.}
}
\description{
print method for class \code{xcorr.output}
}
\keyword{internal}
