% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_table.R
\name{print.selection_table}
\alias{print.selection_table}
\title{print method for class \code{selection_table}}
\usage{
\method{print}{selection_table}(x, no.color = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{selection_table}, generated by \code{\link{selection_table}}.}

\item{...}{further arguments passed to or from other methods. Ignored when printing selection tables.}
}
\description{
print method for class \code{selection_table}
}
\keyword{internal}
