% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec_funs.R
\name{quer_ml}
\alias{quer_ml}
\title{deprecated functions and function names}
\usage{
quer_ml(...)
}
\description{
deprecated functions and function names
}
\details{
functions that have been deprecated or moved to other packages.
}
\keyword{internal}
