% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querxc.R
\name{quer_xc}
\alias{quer_xc}
\title{alternative name for \code{\link{querxc}}}
\usage{
quer_xc(qword, download = FALSE, X = NULL, file.name = c("Genus",
  "Specific_epithet"), parallel = 1, path = NULL, pb = TRUE)
}
\description{
alternative name for \code{\link{querxc}}
}
\details{
see \code{\link{querxc}} for documentation. \code{\link{querxc}} will be deprecated in future versions.
}
\keyword{internal}
