% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_xc.R
\name{querxc}
\alias{querxc}
\title{alternative name for \code{\link{query_xc}}}
\usage{
querxc(...)
}
\description{
alternative name for \code{\link{query_xc}}
}
\details{
Deprecated. See \code{\link{query_xc}}.
}
\keyword{internal}
