% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_xc.R
\name{query_xc}
\alias{query_xc}
\title{Access 'Xeno-Canto' recordings and metadata (deprecated)}
\usage{
query_xc(...)
}
\arguments{
\item{...}{Ignored. Kept so that existing code calling \code{query_xc()} gets an informative warning instead of an error.}
}
\value{
\code{NULL} (invisibly).
}
\description{
\code{query_xc} has been deprecated. Use the package
\href{https://docs.ropensci.org/suwo/}{suwo} instead: \code{suwo::query_xenocanto()}
retrieves 'Xeno-Canto' metadata and \code{suwo::download_media()} downloads the recordings.
}
\details{
This function has been deprecated as access to 'Xeno-Canto' (and other online nature media repositories such as Macaulay Library, iNaturalist, GBIF and WikiAves) is now provided by the package \href{https://docs.ropensci.org/suwo/}{suwo}. Note that 'Xeno-Canto' queries now require an API key (see \code{?suwo::query_xenocanto}). For example:
\preformatted{
# install.packages("suwo")
library(suwo)

# search for metadata
p_anth <- query_xenocanto(species = "Phaethornis anthophilus")

# download recordings
download_media(metadata = p_anth, path = tempdir())
}
}
\seealso{
\code{\link{map_xc}}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
\keyword{internal}
