% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_table.R
\name{rbind.selection_table}
\alias{rbind.selection_table}
\title{rbind method for class \code{selection_table}}
\usage{
\method{rbind}{selection_table}(..., deparse.level = 1)
}
\arguments{
\item{...}{Objects of class \code{selection_table}, generated by \code{\link{selection_table}}.}
}
\description{
rbind method for class \code{selection_table}
}
\keyword{internal}
