% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wave.R
\name{read_wave}
\alias{read_wave}
\title{A wrapper for tuner readWave function internal warbleR function not to be called by users
which the start/end times and frequency range of acoustic signals listed in a data frame can be adjusted.}
\usage{
read_wave(filename, from = 1, to = Inf, units = "seconds",
  header = FALSE, toWaveMC = NULL, ignore.ext.case = TRUE)
}
\description{
A wrapper for tuner readWave function internal warbleR function not to be called by users
which the start/end times and frequency range of acoustic signals listed in a data frame can be adjusted.
}
