% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_resample.R
\name{resample_est}
\alias{resample_est}
\title{alternative name for \code{\link{est_resample}}}
\usage{
resample_est(
  X,
  samp.rate = 44.1,
  bit.depth = 16,
  avoid.clip = TRUE,
  pb = FALSE,
  parallel = 1
)
}
\description{
alternative name for \code{\link{est_resample}}
}
\details{
see \code{\link{est_resample}} for documentation. \code{\link{resample_est}} will be deprecated in future versions.
}
\keyword{internal}
