% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_est.R
\name{resample_est_waves}
\alias{resample_est_waves}
\title{alternative name for \code{\link{resample_est}}}
\usage{
resample_est_waves(
  X,
  samp.rate = 44.1,
  bit.depth = 16,
  avoid.clip = TRUE,
  pb = FALSE,
  parallel = 1
)
}
\description{
alternative name for \code{\link{resample_est}}
}
\details{
see \code{\link{resample_est}} for documentation. \code{\link{resample_est_waves}} will be deprecated in future versions.
}
\keyword{internal}
