% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_silence.R
\name{rm_sil}
\alias{rm_sil}
\title{alternative name for \code{\link{remove_silence}}}
\usage{
rm_sil(
  path = NULL,
  min.sil.dur = 2,
  img = TRUE,
  it = "jpeg",
  flim = NULL,
  files = NULL,
  flist = NULL,
  parallel = 1,
  pb = TRUE,
  downsample = TRUE
)
}
\description{
alternative name for \code{\link{remove_silence}}
}
\details{
see \code{\link{remove_silence}} for documentation. \code{\link{rm_sil}} will be deprecated in future versions.
}
\keyword{internal}
