% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_raven.R
\name{run_raven}
\alias{run_raven}
\title{Open sound files in Raven sound analysis software}
\usage{
run_raven(raven.path = NULL, sound.files = NULL, path = NULL, at.the.time = 10,
import = FALSE, redo = FALSE, ...)
}
\arguments{
\item{raven.path}{A character string indicating the path of the directory in which to look for the raven executable file (where Raven was installed).}

\item{sound.files}{character vector indicating the files that will be analyzed. If  \code{NULL} (default) then Raven will be run without opening any file.}

\item{path}{A character string indicating the path of the directory in which to look for
the sound files. If not provided (default) the function searches into the current working 
directory. Default is \code{NULL}.}

\item{at.the.time}{Numeric vector of length 1 controling how many files will be open in
Raven at the same time. Note that opening too many files at once could make Raven run out
of memory. You need to close Raven every time the batch of files is analyzed, so the next
batch is opened. Default is 10.}

\item{import}{Logical. Controls if the selection tables generated should be returned as a 
data frame into the R environment. This only works if the selections are saved in the 
"Selections" folder in the Raven directory. This argument calls the \code{\link{imp.raven}}
internally. Additional arguments can be passed to \code{\link{imp.raven}} to control the way the data is imported.}

\item{redo}{Logical. Controls whether only the subset of files with no Raven selections (.txt file) in the Raven 'selections' folder
are analyzed. Useful when resuming the analysis. Default is \code{FALSE}.}

\item{...}{Additional arguments to be passed to \code{\link{imp.raven}} for customizing
how selections are imported (ignored if \code{import = FALSE}).}
}
\value{
If \code{import = TRUE} a data frame with the selections produced during the analysis will be return as an data frame. See \code{\link{imp.raven}} for more details on how selections are imported.
}
\description{
\code{run_raven} opens several sound files in Raven sound analysis software
}
\details{
The function runs Raven interactive analysis software (Cornell Lab of Ornithology), opening many files simultaneously. Raven will still run if no sound files are provided (i.e. \code{sound.files = NULL}). At the end of the analysis the data can be automatically imported back into R using the 'import' argument.
}
\examples{
\dontrun{
# First set temporary folder
setwd(tempdir())

# save sound files 
data(list = c("Phae.long1", "Phae.long2"))
writeWave(Phae.long1, "Phae.long1.wav", extensible = FALSE)
writeWave(Phae.long2, "Phae.long2.wav", extensible = FALSE)

raven.path <- "PATH_TO_RAVEN_DIRECTORY_HERE" 

# run function 
run_raven(raven.path = raven.path, sound.files = c("Phae.long1.wav", "Phae.long2.wav"),
 at.the.time = 2, import = T, name.from.file = T, ext.case = "upper", all.data = T)  
 
#getting all the data
rav.dat<-run_raven(all.data = TRUE)
View(rav.dat)

# run function on all the wav files in the working directory 
run_raven(raven.path = raven.path, sound.files = list.files(pattern = "\\.wav$", 
ignore.case = TRUE), at.the.time = 4, import = FALSE)
  
}

}
\author{
Marcelo Araya-Salas (\email{araya-salas@cornell.edu})
}
\seealso{
\code{\link{imp.raven}} \code{\link{imp.syrinx}}
}

