% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.en.ts.R
\name{se_ts}
\alias{se_ts}
\title{alternative name for \code{\link{sp.en.ts}}}
\usage{
se_ts(X, wl = 512, length.out = 20, wn = "hanning", ovlp = 70,
  bp = NULL, threshold = 15, img = TRUE, parallel = 1,
  path = NULL, img.suffix = "sp.en.ts", pb = TRUE,
  clip.edges = FALSE, leglab = "sp.en.ts", sp.en.range = c(2, 10),
  ...)
}
\description{
alternative name for \code{\link{sp.en.ts}}
}
\details{
see \code{\link{sp.en.ts}} for documentation. \code{\link{sp.en.ts}} will be deprecated in future versions.
}
\keyword{internal}
