% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seltailor.R
\name{sel_tailor}
\alias{sel_tailor}
\title{alternative name for \code{\link{seltailor}}}
\usage{
sel_tailor(X = NULL, wl = 512, flim = c(0, 22), wn = "hanning",
  mar = 0.5, osci = TRUE, pal = reverse.gray.colors.2, ovlp = 70,
  auto.next = FALSE, pause = 1, comments = TRUE, path = NULL,
  frange = TRUE, fast.spec = FALSE, ext.window = TRUE, width = 15,
  height = 5, index = NULL, collevels = NULL,
  title = c("sound.files", "selec"), ts.df = NULL, col = "#E37222",
  alpha = 0.7, auto.contour = FALSE, ...)
}
\description{
alternative name for \code{\link{seltailor}}
}
\details{
see \code{\link{seltailor}} for documentation. \code{\link{seltailor}} will be deprecated in future versions.
}
\keyword{internal}
