% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selec.table-data.R
\docType{data}
\name{selec.table}
\alias{selec.table}
\title{Data frame of selections (i.e. selection table).}
\format{A data frame with 11 rows and 6 variables: \describe{ 
 \item{sound.files}{recording names}
 \item{channel}{channel in which signal is found}
 \item{selec}{selection numbers within recording}
 \item{start}{start times of selected signal}
 \item{end}{end times of selected signal}
 \item{bottom.freq}{lower limit of frequency range}
 \item{top.freq}{upper limit of frequency range}
 \item{sel.comment}{selection comments}
 \item{rec.comment}{recording comments}
}}
\source{
Marcelo Araya Salas, warbleR
}
\usage{
data(selec.table)
}
\description{
A data frame containing the start, end, low and hig frequency of 
\emph{Phaethornis longirostris} (Long-billed Hermit) songs from the 
example sound files included in this package.
}
\keyword{datasets}
