% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selec_table-data.R
\docType{data}
\name{selec_table}
\alias{selec_table}
\title{Alternative name for \code{selec.table}}
\format{An object of class \code{data.frame} with 11 rows and 9 columns.}
\usage{
selec_table
}
\description{
\code{selec_table} alternative name for \code{selec.table}. \code{selec.table} will be deprecated in 
future versions.
}
\details{
Simulated coordinated singing events.
}
\keyword{datasets}
