% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig2noise.R
\name{signal_2_noise}
\alias{signal_2_noise}
\title{alternative name for \code{\link{sig2noise}}}
\usage{
signal_2_noise(
  X,
  mar,
  parallel = 1,
  path = NULL,
  pb = TRUE,
  type = 1,
  eq.dur = FALSE,
  in.dB = TRUE,
  before = FALSE,
  lim.dB = TRUE,
  bp = NULL,
  wl = 10
)
}
\description{
alternative name for \code{\link{sig2noise}}
}
\details{
see \code{\link{sig2noise}} for documentation. \code{\link{sig2noise}} will be deprecated in future versions.
}
\keyword{internal}
