% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.coor.sing-data.R
\docType{data}
\name{sim.coor.sing}
\alias{sim.coor.sing}
\title{Simulated coordinated singing events.}
\format{\describe{
  
  \item{sim.coor.sing}{Simulated coordinated singing events that overlap and do not overlap most of the time,
  for use with \code{coor.test}}
  
  }}
\usage{
data(sim.coor.sing)
}
\description{
Simulated coordinated singing events.
}
\keyword{datasets}

