% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_coor_sing-data.R
\docType{data}
\name{sim_coor_sing}
\alias{sim_coor_sing}
\title{Alternative name for \code{sim.coor.song}}
\format{
An object of class \code{data.frame} with 446 rows and 4 columns.
}
\usage{
sim_coor_sing
}
\description{
\code{sim_coor_sing} alternative name for \code{sim.coor.song}. \code{sim.coor.song} will be deprecated in 
future versions.
}
\keyword{datasets}
