% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snr_spectrograms.R
\name{snrspecs}
\alias{snrspecs}
\title{alternative name for \code{\link{snr_spectrograms}}}
\usage{
snrspecs(
  X,
  wl = 512,
  flim = NULL,
  wn = "hanning",
  ovlp = 70,
  inner.mar = c(5, 4, 4, 2),
  outer.mar = c(0, 0, 0, 0),
  picsize = 1,
  res = 100,
  cexlab = 1,
  title = TRUE,
  before = FALSE,
  eq.dur = FALSE,
  propwidth = FALSE,
  xl = 1,
  osci = FALSE,
  gr = FALSE,
  sc = FALSE,
  mar = 0.2,
  snrmar = 0.1,
  it = "jpeg",
  parallel = 1,
  path = NULL,
  pb = TRUE
)
}
\description{
alternative name for \code{\link{snr_spectrograms}}
}
\details{
see \code{\link{snr_spectrograms}} for documentation. \code{\link{snr_spectrograms}} will be deprecated in future versions.
}
\keyword{internal}
