% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_ts.R
\name{sp.en.ts}
\alias{sp.en.ts}
\title{alternative name for \code{\link{entropy_ts}}}
\usage{
sp.en.ts(
  X,
  wl = 512,
  length.out = 20,
  wn = "hanning",
  ovlp = 70,
  bp = "frange",
  threshold = 15,
  img = TRUE,
  parallel = 1,
  path = NULL,
  img.suffix = "entropy_ts",
  pb = TRUE,
  clip.edges = FALSE,
  leglab = "entropy_ts",
  sp.en.range = c(2, 10),
  ...
)
}
\description{
alternative name for \code{\link{entropy_ts}}
}
\details{
see \code{\link{entropy_ts}} for documentation. \code{\link{sp.en.ts}} will be deprecated in future versions.
}
\keyword{internal}
