% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specan.R
\name{spec_an}
\alias{spec_an}
\title{alternative name for \code{\link{specan}}}
\usage{
spec_an(X, bp = "frange", wl = 512, wl.freq = NULL, threshold = 15,
  parallel = 1, fast = TRUE, path = NULL, pb = TRUE, ovlp = 50,
  wn = "hanning", fsmooth = 0.1, harmonicity = FALSE, nharmonics = 3,
  ...)
}
\description{
alternative name for \code{\link{specan}}
}
\details{
see \code{\link{specan}} for documentation. \code{\link{specan}} will be deprecated in future versions.
}
\keyword{internal}
