% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specreator.R
\name{spectrograms}
\alias{spectrograms}
\title{alternative name for \code{\link{specreator}}}
\usage{
spectrograms(X, wl = 512, flim = c(0, 22), wn = "hanning",
  pal = reverse.gray.colors.2, ovlp = 70, inner.mar = c(5, 4, 4, 2),
  outer.mar = c(0, 0, 0, 0), picsize = 1, res = 100, cexlab = 1,
  title = TRUE, propwidth = FALSE, xl = 1, osci = FALSE, gr = FALSE,
  sc = FALSE, line = TRUE, col = adjustcolor("#E37222", 0.6), lty = 3,
  mar = 0.05, it = "jpeg", parallel = 1, path = NULL, pb = TRUE,
  fast.spec = FALSE, by.song = NULL, sel.labels = "selec", ...)
}
\description{
alternative name for \code{\link{specreator}}
}
\details{
see \code{\link{specreator}} for documentation. \code{\link{specreator}} will be deprecated in future versions.
}
\keyword{internal}
