% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sound_files.R
\name{split_wavs}
\alias{split_wavs}
\title{alternative name for \code{\link{split_sound_files}}}
\usage{
split_wavs(
  path = NULL,
  sgmt.dur = 10,
  sgmts = NULL,
  files = NULL,
  parallel = 1,
  pb = TRUE,
  only.sels = FALSE,
  X = NULL
)
}
\description{
alternative name for \code{\link{split_sound_files}}
}
\details{
see \code{\link{split_sound_files}} for documentation. \code{\link{split_wavs}} will be deprecated in future versions.
}
\keyword{internal}
