% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_table.R
\name{[.selection_table}
\alias{[.selection_table}
\title{extract method for class \code{selection_table}}
\usage{
\method{[}{selection_table}(X, i = NULL, j = NULL, drop = TRUE)
}
\arguments{
\item{X}{Object of class \code{selection_table}, generated by \code{\link{selection_table}}.}

\item{i, }{j,    indices specifying elements to extract or replace. Indices are numeric or character vectors or empty 
(missing) or NULL.}
}
\description{
extract method for class \code{selection_table}
}
\keyword{internal}
