% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_diagnostic.R
\name{summarize_diagnostic}
\alias{summarize_diagnostic}
\title{Summarize detection diagnostics}
\usage{
summarize_diagnostic(diagnostic, time.diagnostics = FALSE)
}
\arguments{
\item{diagnostic}{A data frame with the reference selections (start and end of the signals) that will be used to evaluate the performance of the detection, represented by those selections in 'detection'. Must contained at least the following columns: "sound.files", "selec", "start" and "end".}

\item{time.diagnostics}{Logical argument to control if diagnostics related to the duration of the signals ("mean.duration.true.positives", "mean.duration.false.positives", "mean.duration.false.negatives" and "proportional.duration.true.positives") are returned (if \code{TRUE}). Default is \code{FALSE}.}
}
\value{
A data frame, typically the output of a detection optimization function (\code{\link{diagnose_detection}}, \code{\link{optimize_find_peaks}}, \code{\link{optimize_auto_detec}}) including the following detection performance diagnostics:
\itemize{
 \item \code{true.positives}: number of detections that correspond to signals referenced in 'reference'. Matching is defined as some degree of overlap in time. In a perfect detection routine it should be equal to the number of rows in 'reference'. 
 \item \code{false.positives}: number of detections that don't match any of the signals referenced in 'reference'. In a perfect detection routine it should be 0.    
 \item \code{false.negatives}: number of signals in 'reference' that were not detected (not found in 'detection'. In a perfect detection routine it should be 0. 
 \item \code{split.positives}: number of signals referenced in 'reference' that were overlapped by more than 1 detection (i.e. detections that were split). In a perfect detection routine it should be 0.  
 \item \code{mean.duration.true.positives}: mean duration of true positives (in s). Optional.  
 \item \code{mean.duration.false.positives}: mean duration of false positives (in s). Optional.
 \item \code{mean.duration.false.negatives}: mean duration of false negatives (in s). Optional.
 \item \code{proportional.duration.true.positives}: ratio of total duration of true positives to the total duration of signals referenced in 'reference'. In a perfect detection routine it should be 1. Optional.
 \item \code{sensitivity}: Proportion of signals referenced in 'reference' that were detected. In a perfect detection routine it should be 1.
 \item \code{specificity}: Proportion of detections that correspond to signals referenced in 'reference' that were detected. In a perfect detection routine it should be 1.
 }
}
\description{
\code{summarize_diagnostic} summarizes detection diagnostics
}
\details{
The function summarizes a detection diagnostic data frame in which diagnostic parameters are shown split by (typically) a categorical column, usually sound files. This function is used internally by \code{\link{diagnose_detection}}.
}
\examples{
{
# load example selection tables

data(list = c("lbh_selec_reference", "lbh_selec_table"))

# 'lbh_selec_reference' has all signals annotated while 'lbh_selec_table' is missing a few

# run diagnose_detection() by sound file
diag <- diagnose_detection(reference = lbh_selec_reference, 
detection = lbh_selec_table, by.sound.file = TRUE)

# summarize
summarize_diagnostic(diagnostic = diag)

# should be the same as this:
diagnose_detection(reference = lbh_selec_reference, 
detection = lbh_selec_table, by.sound.file = FALSE)
}
}
\references{
{
Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.
}
}
\seealso{
\code{\link{diagnose_detection}}, \code{\link{optimize_find_peaks}}
}
\author{
Marcelo Araya-Salas \email{marcelo.araya@ucr.ac.cr})
}
