% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav_check.R
\name{wav_check}
\alias{wav_check}
\title{Check .wav files}
\usage{
wav_check(X = NULL, path = NULL)
}
\arguments{
\item{X}{Optional. 'selection_table' object or data frame with the following columns: 1) "sound.files": name of the .wav 
files, 2) "sel": number of the selections, 3) "start": start time of selections, 4) "end": 
end time of selections. The output of \code{\link{auto_detec}} can 
also be used as the input data frame. If provided the function also returns the
smallest number of samples from the listed selections, which limits the minimum window 
length (wl argument in other functions) that can be used in batch analyses. 
This could be useful for avoiding errors in downstream functions (e.g. \code{\link{spectro_analysis}}).}

\item{path}{Character string containing the directory path where the sound files are located. 
If \code{NULL} (default) then the current working directory is used.}
}
\value{
If all .wav files are ok, returns message "All files can be read".
  Otherwise returns the names of the corrupted .wav files.
}
\description{
\code{wav_check} checks whether .wav files can be read by subsequent functions.
}
\details{
This function checks if .wav files in the working directory can be read.
Users must set the working directory where they wish to check .wav files beforehand. 
If X is provided it also returns the smallest number of samples from
the selections listed in X (if all files can be read). Note that corrupt files can be
fixed using \code{\link{wav_fix}}) ('sox' must be installed to be able to run this function).
The function is intended for a "quick and dirty" check of the .wav files in a selections data
 frame. For a more thorough analysis see \code{\link{sel_check}}.
}
\examples{
{
# save wav file examples
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "lbh_selec_table"))
writeWave(Phae.long1, file.path(tempdir(), "Phae.long1.wav"))
writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"))
writeWave(Phae.long3, file.path(tempdir(), "Phae.long3.wav"))
writeWave(Phae.long4, file.path(tempdir(), "Phae.long4.wav"))

# without selection data frame
wav_check(path = tempdir())

# without selection data frame
wav_check(X = lbh_selec_table, path = tempdir())
}
}
\references{
{Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.}
}
\seealso{
\code{\link{sel_check}} \code{\link{sel_tailor}}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
