% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration_wavs.R
\name{wav_dur}
\alias{wav_dur}
\title{alternative name for \code{\link{duration_wavs}}}
\usage{
wav_dur(files = NULL, path = NULL)
}
\description{
alternative name for \code{\link{duration_wavs}}
}
\details{
see \code{\link{duration_wavs}} for documentation. \code{\link{wavdur}} will be deprecated in future versions.
}
\keyword{internal}
