% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_wavs.R
\name{wav_info}
\alias{wav_info}
\title{alternative name for \code{\link{info_wavs}}}
\usage{
wav_info(path = NULL, parallel = 1, pb = TRUE)
}
\description{
alternative name for \code{\link{info_wavs}}
}
\details{
see \code{\link{info_wavs}} for documentation. \code{\link{wav_info}} will be deprecated in future versions.
}
\keyword{internal}
