% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_features.R
\name{wpd_features}
\alias{wpd_features}
\title{Measure wavelet packet decomposition features (EXPERIMENTAL)}
\usage{
wpd_features(X, normalize = TRUE, threshold = 1.3, path = NULL, pb = TRUE, parallel = 1)
}
\arguments{
\item{X}{object of class 'selection_table', 'extended_selection_table' or data frame with the following columns: 1) "sound.files": name of the .wav 
files, 2) "sel": number of the selections, 3) "start": start time of selections, 4) "end": 
end time of selections. The ouptut of \code{\link{manualoc}} or \code{\link{autodetec}} can
also be used as the input data frame.}

\item{normalize}{Logical to determine if features are normalized by signal duration.}

\item{threshold}{Amplitude threshold (\%) for frequency range detection. The frequency range (not the cumulative amplitude) is represented as percentage (100\% = highest amplitude). Default is 10.}

\item{path}{Character string containing the directory path where the sound files are located. 
If \code{NULL} (default) then the current working directory is used.}

\item{pb}{Logical argument to control progress bar and messages. Default is \code{TRUE}.}

\item{parallel}{Numeric. Controls whether parallel computing is applied.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}
}
\description{
\code{wpd_features} Measure wavelet packet decomposition features.
}
\details{
Measures wavelet packet decomposition features. STLL IN DEVELOPMENT. USE IT UNDER YOUR OWN RISK.
}
\examples{
{
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "lbh_selec_table"))
writeWave(Phae.long1, file.path(tempdir(), "Phae.long1.wav"))
writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"))
writeWave(Phae.long3, file.path(tempdir(), "Phae.long3.wav"))
writeWave(Phae.long4, file.path(tempdir(), "Phae.long4.wav"))

# not normalize
wpd_features(lbh_selec_table, threshold = 1.3, nor = FALSE)
}

}
\references{
{
Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.
Selin A., J. Turunen, and J. T. Tanttu, 2007. Wavelets in recognition of bird sounds. EURASIP Journal on Advances in Signal Processing.
}
}
\seealso{
\code{\link{mfcc_stats}}, \code{\link{mfcc_stats}}
}
\author{
Marcelo Araya-Salas (\email{marceloa27@gmail.com})
}
