% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcorr.R
\name{x_corr}
\alias{x_corr}
\title{alternative name for \code{\link{xcorr}}}
\usage{
x_corr(X = NULL, wl = 512, frange = NULL, ovlp = 90, dens = 0.9,
  bp = NULL, wn = "hanning", cor.method = "pearson", parallel = 1,
  path = NULL, pb = TRUE, na.rm = FALSE, dfrange = FALSE,
  cor.mat = TRUE)
}
\description{
alternative name for \code{\link{xcorr}}
}
\details{
see \code{\link{xcorr}} for documentation. \code{\link{xcorr}} will be deprecated in future versions.
}
\keyword{internal}
