% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmaps.R
\name{xc_maps}
\alias{xc_maps}
\title{alternative name for \code{\link{xcmaps}}}
\usage{
xc_maps(X, img = TRUE, it = "jpeg", res = 100, labels = FALSE, path = NULL)
}
\description{
alternative name for \code{\link{xcmaps}}
}
\details{
see \code{\link{xcmaps}} for documentation. \code{\link{xcmaps}} will be deprecated in future versions.
}
\keyword{internal}
