% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcorr.graph.R
\name{xcorr_graph}
\alias{xcorr_graph}
\title{alternative name for \code{\link{xcorr.graph}}}
\usage{
xcorr_graph(X, cex.cor = 1, cex.lab = 1, cex.axis.lab = 1,
  rel.cex = FALSE, labs = NULL)
}
\description{
alternative name for \code{\link{xcorr.graph}}
}
\details{
see \code{\link{xcorr.graph}} for documentation. \code{\link{xcorr.graph}} will be deprecated in future versions.
}
\keyword{internal}
