% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Arre.aura-data.R
\docType{data}
\name{Arre.aura}
\alias{Arre.aura}
\title{Acoustic recording of \emph{Arremon aurantiirostris} (Orange-billed Sparrow).}
\format{One .wav file: \describe{

\item{Arre.aura}{\emph{Arremon aurantiirostris} recording}

 }}
\source{
\url{http://www.xeno-canto.org/contributor/EMCWQLLKEW}
}
\usage{
data(Arre.aura)
}
\description{
Acoustic recording of \emph{Arremon aurantiirostris} (Orange-billed Sparrow).
}
\keyword{datasets}

