% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkwavs.R
\name{check_wavs}
\alias{check_wavs}
\title{alternative name for \code{\link{checkwavs}}}
\usage{
check_wavs(X = NULL, path = NULL)
}
\description{
alternative name for \code{\link{checkwavs}}
}
\details{
see \code{\link{checkwavs}} for documentation. \code{\link{checkwavs}} will be deprecated in future versions.
}
\keyword{internal}
