% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sound_files.R
\name{check_wavs}
\alias{check_wavs}
\title{alternative name for \code{\link{check_sound_files}}}
\usage{
check_wavs(X = NULL, path = NULL)
}
\description{
alternative name for \code{\link{check_sound_files}}
}
\details{
see \code{\link{check_sound_files}} for documentation. \code{\link{check_wavs}} will be deprecated in future versions.
}
\keyword{internal}
