% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav_check.R
\name{checkwavs}
\alias{checkwavs}
\title{alternative name for \code{\link{wav_check}}}
\usage{
checkwavs(X = NULL, path = NULL)
}
\description{
alternative name for \code{\link{wav_check}}
}
\details{
see \code{\link{wav_check}} for documentation. \code{\link{checkwavs}} will be deprecated in future versions.
}
\keyword{internal}
