% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkwavs_1.0.R
\name{checkwavs}
\alias{checkwavs}
\title{Check .wav files}
\usage{
checkwavs()
}
\value{
If all .wav files are ok, returns message "All files are ok!".
  Otherwise returns "These file(s) cannot be read" message with names of the
  corrupted .wav files.
}
\description{
\code{checkwavs} checks whether .wav files can be read by subsequent functions.
}
\details{
This function has no associated arguments, although the user must
  set the working directory where they wish to check .wav files beforehand.
}
\examples{
data(list = c("Phae.long1", "Phae.long2", "Phae.long3"))
writeWave(Phae.long1,"Phae.long1.wav")
writeWave(Phae.long2,"Phae.long2.wav")
writeWave(Phae.long3,"Phae.long3.wav")
checkwavs()
}

