% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coor.sing-data.R
\docType{data}
\name{coor.sing}
\alias{coor.sing}
\title{Selections of simulated interactive singing events. The simulated events use the mean and standard
deviation of real lekking \emph{Phaethornis longirostris} (Long-billed Hermit hummingbird) songs and
intervals between songs. Three events are simulated: synchronized overlaping signals (ovlp),
synchronized not overlapping signals (no.ovlp) and not synchronized signals (no.sync). For the not
synchronized simulated event one "birds" sings at the solo singing rate while the
second one sings at the duetting rate (lower singing rate).
.}
\format{\describe{

  \item{coor.sing}{Selections of coordinated singing}

  }}
\usage{
data(coor.sing)
}
\description{
Selections of simulated interactive singing events. The simulated events use the mean and standard
deviation of real lekking \emph{Phaethornis longirostris} (Long-billed Hermit hummingbird) songs and
intervals between songs. Three events are simulated: synchronized overlaping signals (ovlp),
synchronized not overlapping signals (no.ovlp) and not synchronized signals (no.sync). For the not
synchronized simulated event one "birds" sings at the solo singing rate while the
second one sings at the duetting rate (lower singing rate).
.
}
\keyword{datasets}

