% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coor.test.R
\name{coor.test}
\alias{coor.test}
\title{Monte Carlo randomization test to assess the statistical significance of singing coordination}
\usage{
coor.test(X, iterations = 1000, less.than.chance = TRUE)
}
\arguments{
\item{X}{Data frame containing columns for singing event (sing.event),
individual (indiv), and start and end time of signal (start and end).}

\item{iterations}{number of iterations for shuffling and calculation of the expected number of overlaps.}

\item{less.than.chance}{Logical. If \code{TRUE} the test evaluates whether overlaps occur less often than expected by chance.
If \code{FALSE} the opposite pattern is evaluted (whether overlaps occur more often than expected by chance).
Default is  \code{TRUE}.}
}
\value{
A data frame with the observed number of overlaps (obs.overlaps), mean number of overlaps expected by chance,
and p value.
}
\description{
\code{coor.test} assess the statistical significance of singing coordination
}
\details{
This function calculates the probability of finding and equal or lower number
(of higher if les.than.chance is \code{TRUE}) of song overlaps in a coordinated singing event.
The function shuffles the sequences of signals and silence-between-signals for both individuals to produce
a null distribution of expected number of overlaps by chance.  The observed number of overlaps is compared to this
expected values. The p-values are calculated as the proportion of random expected values that were lower (or higher)
than the observed value. The function runs a separated test for each singing event in the input data frame.
}
\examples{
\dontrun{
#load data
data(coor.sing)

# testing if coordination happens less than expected by chance
coor.test(coor.sing, iterations = 1000, less.than.chance = T)

# testing if coordination happens more than expected by chance
coor.test(coor.sing, iterations = 1000, less.than.chance = F)
}
}
\author{
Marcelo Araya-Salas (\url{http://marceloarayasalas.weebly.com/})
}

