% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_coordination.R
\name{coor.test}
\alias{coor.test}
\title{alternative name for \code{\link{test_coordination}}}
\usage{
coor.test(
  X = NULL,
  iterations = 1000,
  ovlp.method = "count",
  randomization = "keep.gaps",
  less.than.chance = TRUE,
  parallel = 1,
  pb = TRUE,
  rm.incomp = FALSE,
  cutoff = 2,
  rm.solo = FALSE
)
}
\description{
alternative name for \code{\link{test_coordination}}
}
\details{
see \code{\link{test_coordination}} for documentation. \code{\link{coor.test}} will be deprecated in future versions.
}
\keyword{internal}
