% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration_sound_files.R
\name{duration_wavs}
\alias{duration_wavs}
\title{alternative name for \code{\link{duration_sound_files}}}
\usage{
duration_wavs(
  files = NULL,
  path = NULL,
  skip.error = FALSE,
  file.format = "\\\\.wav$|\\\\.wac$|\\\\.mp3$|\\\\.flac$"
)
}
\description{
alternative name for \code{\link{duration_sound_files}}
}
\details{
see \code{\link{duration_sound_files}} for documentation. \code{\link{duration_wavs}} will be deprecated in future versions.
}
\keyword{internal}
