% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec_funs.R
\name{ffDTW}
\alias{ffDTW}
\title{Deprecated functions and function names}
\usage{
ffDTW(...)
}
\description{
Deprecated functions and function names
}
\details{
functions that have been deprecated or moved to other packages.
}
\keyword{internal}
