% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixwavs.R
\name{fix_wavs}
\alias{fix_wavs}
\title{alternative name for \code{\link{fixwavs}}}
\usage{
fix_wavs(checksels = NULL, files = NULL, samp.rate = NULL,
  bit.rate = NULL, path = NULL, ...)
}
\description{
alternative name for \code{\link{fixwavs}}
}
\details{
see \code{\link{fixwavs}} for documentation. \code{\link{fixwavs}} will be deprecated in future versions.
}
\keyword{internal}
