% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_wavs.R
\name{fixwavs}
\alias{fixwavs}
\title{alternative name for \code{\link{fix_wavs}}}
\usage{
fixwavs(
  checksels = NULL,
  files = NULL,
  samp.rate = NULL,
  bit.depth = NULL,
  path = NULL,
  mono = FALSE
)
}
\description{
alternative name for \code{\link{fix_wavs}}
}
\details{
see \code{\link{fix_wavs}} for documentation. \code{\link{fixwavs}} will be deprecated in future versions.
}
\keyword{internal}
