% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_range_detec.R
\name{frange.detec}
\alias{frange.detec}
\title{alternative name for \code{\link{freq_range_detec}}}
\usage{
frange.detec(
  wave,
  wl = 512,
  fsmooth = 0.1,
  threshold = 10,
  dB.threshold = NULL,
  wn = "hanning",
  flim = NULL,
  bp = NULL,
  fast.spec = FALSE,
  ovlp = 50,
  pal = reverse.gray.colors.2,
  widths = c(2, 1),
  main = NULL,
  plot = TRUE,
  all.detec = FALSE
)
}
\description{
alternative name for \code{\link{freq_range_detec}}
}
\details{
see \code{\link{freq_range_detec}} for documentation. \code{\link{frange.detec}} will be deprecated in future versions.
}
\keyword{internal}
