\name{warbleR-internals}
\alias{spectroW}
\alias{spectro2}

\title{
\code{warbleR} Internal Functions
}
\description{
These functions are used by other functions in the warbleR package, but are not intended to be called by users. Both are modified versions of \code{\link[seewave]{spectro}}.
The functions are:
\code{spectro2}
\code{spectro2}
}

\keyword{internal}

\note{
  These functions are not to be called by the user.
}