\name{warbleR-internals}
\alias{filled.contour.color.INTFUN}
\alias{filled.contour.INTFUN}
\alias{frd.INTFUN}
\alias{frd.plot.INTFUN}
\alias{spectro.rectw.INTFUN.2}
\alias{spectro.rectw.INTFUN}
\alias{boxw.INTFUN}
\title{
\code{warbleR} Internal Functions
}
\description{
These functions are used by other functions in the warbleR package, but are not intended to be called by users. 

These 2 functions are modified versions of \code{\link[seewave]{spectro}}.
The functions are:
\code{spectro.rectw.INTFUN.2}
\code{spectro.rectw.INTFUN}

These 2 functions are modified versions of \code{\link[seewave]{filled.contour.modif2}}.
\code{\link[seewave]{spectro}}.
The functions are:
\code{filled.contour.color.INTFUN}
\code{filled.contour.INTFUN}

\code{boxw.INTFUN} to create boxes around spectrograms in \code{\link[warbleR]{catalog}}

\code{rectw} is a modified version of \code{\link[graphics]{rect}} to add density (cross-hatching lines)

These 2 functions are used for frequency range detection:
\code{frd.INTFUN}
\code{frd.plot.INTFUN}

}

\keyword{internal}

\note{
  These functions are not to be called by the user.
}