% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbh_selec_table-data.R
\docType{data}
\name{lbh_selec_table}
\alias{lbh_selec_table}
\title{Example data frame of selections (i.e. selection table).}
\format{
A data frame with 11 rows and 6 variables: \describe{ 
 \item{sound.files}{recording names}
 \item{channel}{channel in which signal is found}
 \item{selec}{selection numbers within recording}
 \item{start}{start times of selected signal}
 \item{end}{end times of selected signal}
 \item{bottom.freq}{lower limit of frequency range}
 \item{top.freq}{upper limit of frequency range}
}
}
\source{
Marcelo Araya Salas, warbleR
}
\usage{
data(lbh_selec_table)
}
\description{
\code{lbh_selec_table} alternative name for \code{selec.table}. \code{selec.table} will be deprecated in 
future versions.
}
\details{
A data frame containing the start, end, low and high frequency of 
\emph{Phaethornis longirostris} (Long-billed Hermit) songs from the 
example sound files included in this package. Same data than 'selec_table'.
'selec_table' will be removed in future package version.
}
\keyword{datasets}
