% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_spectrograms.R
\name{lspec}
\alias{lspec}
\title{alternative name for \code{\link{full_spectrograms}}}
\usage{
lspec(
  X = NULL,
  flim = NULL,
  sxrow = 5,
  rows = 10,
  collevels = seq(-40, 0, 1),
  ovlp = 50,
  parallel = 1,
  wl = 512,
  gr = FALSE,
  pal = reverse.gray.colors.2,
  cex = 1,
  it = "jpeg",
  flist = NULL,
  overwrite = TRUE,
  path = NULL,
  pb = TRUE,
  fast.spec = FALSE,
  labels = "selec",
  horizontal = FALSE,
  song = NULL,
  suffix = NULL,
  ...
)
}
\description{
alternative name for \code{\link{full_spectrograms}}
}
\details{
see \code{\link{full_spectrograms}} for documentation. \code{\link{lspec}} will be deprecated in future versions.
}
\keyword{internal}
