% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xc_maps.R
\name{map_xc}
\alias{map_xc}
\title{Maps of 'Xeno-Canto' recordings by species}
\usage{
map_xc(
  X,
  img = TRUE,
  it = "jpeg",
  res = 100,
  labels = FALSE,
  path = NULL,
  leaflet.map = FALSE,
  leaflet.cluster = FALSE
)
}
\arguments{
\item{X}{Data frame output from \code{\link{query_xc}}.}

\item{img}{A logical argument specifying whether an image file of each species
map should be returned, default is \code{TRUE}.}

\item{it}{A character vector of length 1 giving the image type to be used. Currently only
"tiff" and "jpeg" are admitted. Default is "jpeg".}

\item{res}{Numeric argument of length 1. Controls image resolution.
Default is 100 (faster) although 300 - 400 is recommended for publication/
presentation quality.}

\item{labels}{A logical argument defining whether dots depicting recording locations are labeled.
If \code{TRUE} then the Recording_ID is used as label.}

\item{path}{Character string with the directory path where the image files will be saved.
If \code{NULL} (default) then the current working directory is used.
Ignored if \code{img = FALSE}.}

\item{leaflet.map}{Logical to control whether the package 'leaflet' is used for displaying the maps. 'leaflet' maps are interactive and display information about recordings and links to the Xeno-Canto website. If \code{TRUE} a single map is displayed regardless of the number of species and all other image related arguments are ignored. Default is \code{FALSE}. The hovering label shows the species scientific name (or the subspecies if only 1 species is present in 'X'). Note that colors will be recycled if more after 18 species (or subspecies).}

\item{leaflet.cluster}{Logical to control if icons are clustered by locality (as in Xeno-Canto maps). Default is \code{FALSE}.}
}
\value{
A map of 'Xeno-Canto' recordings per species (image file), or a faceted
  plot of species map(s) in the active graphic device.
}
\description{
\code{map_xc} creates maps to visualize the geographic spread of 'Xeno-Canto'
  recordings.
}
\details{
This function creates maps for visualizing the geographic spread of recordings from the open-access
online repository \href{https://www.xeno-canto.org/}{Xeno-Canto}. The function takes the output of
\code{\link{query_xc}} as input. Maps can be displayed in the graphic device (or Viewer if 'leaflet.map = TRUE') or saved as images in the
working directory. Note that only recordings with geographic coordinates are displayed.
}
\examples{
\dontrun{
# search in xeno-canto
X <- query_xc("Phaethornis anthophilus", download = FALSE)

# create image in R graphic device
map_xc(X, img = FALSE)

# create leaflet map
map_xc(X, leaflet.map = TRUE)
}

}
\references{
Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr}) and Grace Smith Vidaurre
}
