% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mp32wav_1.0.R
\name{mp32wav}
\alias{mp32wav}
\title{Convert .mp3 files to .wav}
\usage{
mp32wav()
}
\value{
.wav files saved in the working directory with same name as original mp3 files.
}
\description{
\code{mp32wav} convert several .mp3 files in working directory to .wav format
}
\details{
Currently only working in windows OS. Function to read mp3 from tuneR package crashes in other operating systems.
}
\examples{
\dontrun{
# Need to have at least 1 mp3 file in working directory
mp32wav()
}
}
\author{
Marcelo Araya-Salas (http://marceloarayasalas.weebly.com/) and Grace Smith Vidaurre
}

