% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mp32wav.R
\name{mp32wav}
\alias{mp32wav}
\title{Convert .mp3 files to .wav}
\usage{
mp32wav()
}
\value{
.wav files saved in the working directory with same name as original mp3 files.
}
\description{
\code{mp32wav} convert several .mp3 files in working directory to .wav format
}
\details{
convert .mp3 files in working directory to .wav format. Function to read mp3 from tuneR package sometimes crashes.
Apparently it Will be fixed in the next version of tuneR.
}
\examples{
\dontrun{
# First create empty folder
dir.create(file.path(getwd(),"temporal"))
setwd(file.path(getwd(),"temporal"))

#Then download mp3 files from xeno-canto
querxc(qword = "Phaethornis aethopygus", download = T)

# Convert to wa
mp32wav()

#check this folder!!
getwd()

unlink(getwd(),recursive = T)
}
}
\author{
Marcelo Araya-Salas (\url{http://marceloarayasalas.weebly.com/}) and Grace Smith Vidaurre
}

