% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_function_names.R
\docType{data}
\name{new_function_names}
\alias{new_function_names}
\title{Data frame detailing function name changes}
\format{
An object of class \code{data.frame} with 31 rows and 2 columns.
}
\usage{
data(new_function_names)
}
\description{
A data frame containing the old and new names for \code{\link{warbleR}} functions
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
\keyword{datasets}
