% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Phae.cuvi-data.R
\docType{data}
\name{Phae.cuvi}
\alias{Phae.cuvi}
\title{Acoustic recording of \emph{Phaeochroa cuvierii} (Scaly-breasted Hummingbird).}
\format{One .wav file: \describe{

\item{Phae.cuvi}{\emph{Phaeochroa cuvierii} recording}

 }}
\usage{
data(Phae.cuvi)
}
\description{
Acoustic recording of \emph{Phaeochroa cuvierii} (Scaly-breasted Hummingbird).
}
\keyword{datasets}

