% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Phae.long1-data.R
\docType{data}
\name{Phae.long1}
\alias{Phae.long1}
\title{Acoustic recording #1 of \emph{Phaethornis longirostris} (Long-billed Hermit).}
\format{One .wav file: \describe{

  \item{Phae.long1}{\emph{Phaethornis longirostris} #1 recording}

  }}
\source{
\url{http://www.xeno-canto.org/contributor/EMCWQLLKEW}
}
\usage{
data(Phae.long1)
}
\description{
Acoustic recording #1 of \emph{Phaethornis longirostris} (Long-billed Hermit).
}
\keyword{datasets}

