% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phae.long1-data.R
\docType{data}
\name{Phae.long1}
\alias{Phae.long1}
\title{Audio recording #1 of \emph{Phaethornis longirostris}}
\format{One .wav file: \describe{
  
  \item{Phae.long1}{\emph{Phaethornis longirostris} #1 recording}
  
  }}
\source{
\url{http://www.xeno-canto.org/contributor/EMCWQLLKEW}
}
\usage{
data(Phae.long1)
}
\description{
Audio recording #1 of \emph{Phaethornis longirostris} (Long-billed Hermit).
}
\keyword{datasets}
