% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phae.long2-data.R
\docType{data}
\name{Phae.long2}
\alias{Phae.long2}
\title{Audio recording #2 of \emph{Phaethornis longirostris}}
\format{One .wav file: \describe{
  
  \item{Phae.long2}{\emph{Phaethornis longirostris} #2 recording}
  
  }}
\source{
\url{http://www.xeno-canto.org/contributor/EMCWQLLKEW}
}
\usage{
data(Phae.long2)
}
\description{
Audio recording #2 of \emph{Phaethornis longirostris}  (Long-billed Hermit).
}
\keyword{datasets}
