% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Phae.long3-data.R
\docType{data}
\name{Phae.long3}
\alias{Phae.long3}
\title{Acoustic recording #3 of \emph{Phaethornis longirostris} (Long-billed Hermit).}
\format{One .wav file: \describe{

  \item{Phae.long3}{\emph{Phaethornis longirostris} #3 recording}

  }}
\source{
\url{http://www.xeno-canto.org/contributor/EMCWQLLKEW}
}
\usage{
data(Phae.long3)
}
\description{
Acoustic recording #3 of \emph{Phaethornis longirostris} (Long-billed Hermit).
}
\keyword{datasets}

