% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phae.long4-data.R
\docType{data}
\name{Phae.long4}
\alias{Phae.long4}
\title{Audio recording #1 of \emph{Phaethornis longirostris}}
\format{One .wav file: \describe{
  
  \item{Phae.long4}{\emph{Phaethornis longirostris} #4 recording}
  
  }}
\source{
\url{http://www.xeno-canto.org/contributor/EMCWQLLKEW}
}
\usage{
data(Phae.long4)
}
\description{
Audio recording #4 of \emph{Phaethornis longirostris} (Long-billed Hermit).
}
\keyword{datasets}
