% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quer_ml.R
\name{quer_ml}
\alias{quer_ml}
\title{Access Macaulay Library media metadata}
\usage{
quer_ml(qword, download = FALSE, X = NULL, file.name = "sciName", 
parallel = 1, path = NULL, pb = TRUE)
}
\arguments{
\item{qword}{Character vector of length one indicating the genus, or genus and
species, to query Macaulay Library database. For example, \emph{Phaethornis} or \emph{Phaethornis longirostris}. More complex queries will be available in the 
future using search terms that follow the Macaulay Library advance query
 syntax.}

\item{download}{Logical argument. If \code{FALSE} only the recording file names and
associated metadata are downloaded. If \code{TRUE}, recordings are also downloaded to the working
directory as .mp3 files. Default is \code{FALSE}. Note that if the recording is already in the 
working directory (as when the downloading process has been interrupted) it will be skipped. 
Hence, resuming downloading processes will not start from scratch.}

\item{X}{Data frame with a 'catalogId' column and any other column listed in the file.name argument. Only the recordings listed in the data frame 
will be download (\code{download} argument is automatically set to \code{TRUE}). This can be used to select
the recordings to be downloaded based on their attributes.}

\item{file.name}{Character vector indicating the tags (or column names) to be included in the sound file names (if download = \code{TRUE}). Several tags can be included. If \code{NULL} only the Macaulay Library catalog identification number ("catalogId") is used. Default is "sciName".
Note that "catalogId" is always used (whether or not is listed by users) to avoid duplicated names.}

\item{parallel}{Numeric. Controls whether parallel computing is applied when downloading mp3 files.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing). Applied both when getting metadata and downloading files.}

\item{path}{Character string containing the directory path where the sound files will be saved. 
If \code{NULL} (default) then the current working directory is used.}

\item{pb}{Logical argument to control progress bar. Default is \code{TRUE}.}
}
\value{
If X is not provided the function returns a data frame with the following recording information: recording ID, Genus, Specific epithet, Subspecies, English name, Recordist, Country, Locality, Latitude, Longitude, Vocalization type, Audio file, License, URL, Quality, Time, Date. Sound files in .mp3 format are downloaded into the working directory if download = \code{TRUE} or if X is provided; a column indicating the  names of the downloaded files is included in the output data frame.
}
\description{
\code{quer_ml} downloads recordings and metadata from Macaulay Library database (\url{http://macaulaylibrary.org/}).
}
\details{
This function queries for animal sound recordings at the Macaulay Library, Cornell Lab of Ornithology (\url{http://macaulaylibrary.org/}). It can return recordings metadata
or download the associated sound files.
 Files are double-checked after downloading and "empty" files are re-downloaded.
 File downloading process can be interrupted and resume later as long as the 
 working directory is the same. Note that only the first 100 results can be 
 retreived at this time.
}
\examples{
\dontrun{
# Set temporary working directory
# setwd(tempdir())

# search without downloading
df1 <- quer_ml(qword = 'Phaethornis squalidus', download = FALSE)
View(df1)

# downloading files
quer_ml(qword = 'Phaethornis squalidus', download = TRUE)

# check folder
open_wd()
}
}
\seealso{
\code{\link{quer_xc}}
}
\author{
Marcelo Araya-Salas (\email{araya-salas@cornell.edu})
}
