% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xc_query.R
\name{querxc}
\alias{querxc}
\title{alternative name for \code{\link{xc_query}}}
\usage{
querxc(
  qword,
  download = FALSE,
  X = NULL,
  file.name = c("Genus", "Specific_epithet"),
  parallel = 1,
  path = NULL,
  pb = TRUE
)
}
\description{
alternative name for \code{\link{xc_query}}
}
\details{
see \code{\link{xc_query}} for documentation. \code{\link{querxc}} will be deprecated in future versions.
}
\keyword{internal}
