% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/querxc_1.1.R
\name{querxc}
\alias{querxc}
\title{Access Xeno Canto recordings and metadata}
\usage{
querxc(qword, download = FALSE)
}
\arguments{
\item{qword}{Character vector of length one or two (genus or genus and
species). Used to query Xeno Canto database.}

\item{download}{Logical argument, downloads recording file names and
associated metadata if FALSE. If TRUE, recordings are downloaded to working
directory as .mp3 files. Default is FALSE.}
}
\value{
Data frame with recording information, or .mp3 files.
}
\description{
\code{querxc} downloads recordings and metadata from Xeno Canto.
}
\examples{
\dontrun{
X <- querxc("Phaethornis anthophilus", download = FALSE)
querxc("Phaethornis anthophilus", download = TRUE)
}
}
\author{
Marcelo Araya-Salas http://marceloarayasalas.weebly.com/ and Hua Zhong
}

