% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selec.table-data.R
\docType{data}
\name{selec.table}
\alias{selec.table}
\title{Alternative name for \code{selec.table}}
\format{
An object of class \code{data.frame} with 11 rows and 9 columns.
}
\source{
Marcelo Araya Salas, warbleR
}
\usage{
selec.table
}
\description{
\code{selec.table} alternative name for \code{lbh_selec_table}. \code{selec.table} will be deprecated in 
future versions.
}
\details{
Simulated coordinated singing events.
}
\keyword{datasets}
