% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_table.R
\name{selection_table}
\alias{selection_table}
\title{Create 'selection_table' class objects}
\usage{
selection_table(X, max.dur = 10, path = NULL, whole.recs = FALSE,
extended = FALSE, confirm.extended = TRUE, mar = 0.05, by.song = NULL, ...)
}
\arguments{
\item{X}{data frame with the following columns: 1) "sound.files": name of the .wav 
files, 2) "sel": number of the selections, 3) "start": start time of selections, 4) "end": 
end time of selections. Columns for 'top.freq', 'bottom.freq' and 'channel' are optional.  Alternatively, a 'selection_table' class object can be input to double check selections. 
The ouptut of \code{\link{manualoc}} or \code{\link{autodetec}} can 
be used as the input object for other \code{\link{warbleR}} functions.}

\item{max.dur}{the maximum duration of expected for a selection  (ie. end - start).}

\item{path}{Character string containing the directory path where the sound files are located. 
If \code{NULL} (default) then the current working directory is used.}

\item{whole.recs}{Logical. If \code{TRUE} the function will create a selection 
table for all sound files in the working directory (or "path") with `start = 0` 
and `end = wavdur()`. Default is if \code{FALSE}.}

\item{extended}{Logical. If \code{TRUE}, the selections are 
included as an additional attribute ('wave.objects') to the data set, which now will
be called a 'extended selection 
table'. This can largely faciliate the storing and sharing of (bio)acoustic data
as the extended selection tables will contain all the necessary information for 
running acoustic analysis. The output object will also belong to the 'extended_selection_table' 
class (in addition to the 'data.frame' and 'selection_table' classes). Default 
is if \code{FALSE}. An 'extended selection table won't be created if there is any issue with the selection. See 'details'.}

\item{confirm.extended}{Logical. If \code{TRUE} then the size of the 'extended_selection_table' 
will be estimated and the user will be asked for confirmation (in the console)
before proceeding. Ignored if 'extended' is \code{FALSE}. This is used to prevent
generating objects too big to be dealth with by R. See 'details' for more information about extended selection table size.}

\item{mar}{Numeric vector of length 1 specifying the margins (in seconds) 
adjacent to the start and end points of the selections when creating extended 
selection tables. Default is 0.05. Ignored if 'extended' is \code{FALSE}.}

\item{by.song}{Character string with the column name containing song labels. If provided a wave object containing for 
all selection belonging to a single song would be saved in the extended selection table (hence only applicable for 
extended selection tables). Default is \code{FALSE}. Ignored if \code{extended = FALSE}. Note that the function assumes
that song labels are not repeated within a sound file. If \code{NULL} (default), wave objects are created for each selection.}

\item{...}{Additional arguments to be passed to \code{\link{checksels}} for customizing
checking routine.}
}
\value{
An object of class selection_table which includes the original data frame plus the following additional attributes:
\itemize{
   \item 1) A data frame with the output of \code{\link{checksels}} run on the input data frame. If a extended selection table is created it will also include the original values in the input data frame for each selections. This are used by downstream warbleR functions to improve efficiency and avoid
errors due to missing or mislabeled data, or selection out of the ranges of the original sound files. 
   \item 2) A list indicating if the selection table has been created by song (see 'by.song argument).
   \item 3) If a extended selection table is created a list containing the wave objects for each selection (or song if 'by.song').
   }
}
\description{
\code{selection_table} converts data frames into an object of class selection_table.
}
\details{
This function creates and object of class 'selection_table' or 'extended_selection_table' (if \code{extended = TRUE}, see below). The function checks: 
\itemize{
   \item 1) if the selections listed in the data frame correspond to .wav files
in the working directory
   \item 2) if the sound files can be read and if so, 
   \item 3) if the start and end time of the selections are found within the duration of the sound files
   }
Note that the sound files should be in the working directory (or the directory provided in 'path').
This is useful for avoiding errors in dowstream functions (e.g. \code{\link{specan}}, \code{\link{xcorr}}, \code{\link{catalog}}, \code{\link{dfDTW}}). Note also that corrupt files can be
fixed using \code{\link{fixwavs}} ('sox' must be installed to be able to run this function).
The 'selection_table' class can be input in subsequent functions. 

When \code{extended = TRUE} the function will generate an object of class 'extended_selection_table' which 
will also contain the wave objects for each of the selections in the data frame.
This transforms selection talbe in stand alone objects as they no longer need the original
sound files to run acoustic analysis. Extended selection table size will be a 
function of the number of selections \code{nrow(X)}, sampling rate, selection 
duration and margin duration. As a guide, a selection table
with 1000 selections similar to the ones in 'selec.table' (mean duration ~0.15
 seconds) at 22.5 kHz sampling rate and the default margin (\code{mar = 0.05}) 
 will generate a extended selection table of ~23 MB (~230 MB for a 10000 rows 
 selection table). You can check the size of the output extended selection table
 with the \code{\link[utils]{object.size}} function.
}
\examples{
{
# First set temporary folder
 setwd(tempdir())

data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "selec.table"))
writeWave(Phae.long1,"Phae.long1.wav")
writeWave(Phae.long2,"Phae.long2.wav")
writeWave(Phae.long3,"Phae.long3.wav")
writeWave(Phae.long4,"Phae.long4.wav")

# make selection table
st <- selection_table(X = selec.table)

is_selection_table(st)

#' # make extended selection table
st <- selection_table(X = selec.table, extended = TRUE, confirm.extended = FALSE)

is_extended_selection_table(st)

### make extended selection by song  
# create a song variable
selec.table$song <- as.numeric(selec.table$sound.files)

st <- selection_table(X = selec.table, extended = TRUE, confirm.extended = FALSE, by.song = "song")
}
}
\seealso{
\code{\link{checkwavs}}
}
\author{
Marcelo Araya-Salas (\email{araya-salas@cornell.edu})
}
