% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.coor.sing-data.R
\docType{data}
\name{sim.coor.sing}
\alias{sim.coor.sing}
\title{Simulated coordinated singing events.}
\format{
\describe{
  
  \item{sim.coor.sing}{Simulated coordinated singing events that overlap and do not overlap most of the time,
  for use with \code{test_coordination}}
  
  }
}
\usage{
data(sim.coor.sing)
}
\description{
\code{sim.coor.sing} are selections of simulated interactive singing events.The simulated events use the mean 
and standard deviation of real lekking \emph{Phaethornis longirostris} (Long-billed 
Hermit hummingbird) songs and intervals between songs (e.i gaps). Three events are simulated: overlapping signals (ovlp), 
alternating signals (altern) and non-synchronized signals (uncoor). Will be replaced by 'sim_coor_sing' in future package versions.
}
\keyword{datasets}
