% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectro_tweak.R
\name{spec_param}
\alias{spec_param}
\title{alternative name for \code{\link{spectro_tweak}}}
\usage{
spec_param(
  X,
  length.out = 5,
  ovlp = 90,
  wl = c(100, 1000),
  wn = "hanning",
  collev.min = -40,
  pal = "reverse.gray.colors.2",
  path = NULL,
  rm.axes = TRUE,
  ...
)
}
\description{
alternative name for \code{\link{spectro_tweak}}
}
\details{
see \code{\link{spectro_tweak}} for documentation. \code{\link{spec_param}} will be deprecated in future versions.
}
\keyword{internal}
