% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specan.R
\name{specan}
\alias{specan}
\title{Measure acoustic parameters in batches of sound files}
\usage{
specan(X, bp = c(0,22), wl = 512, wl.freq = NULL, threshold = 15,
 parallel = 1, fast = TRUE, path = NULL, pb = TRUE, ovlp = 50, ff.method = "seewave", 
wn = "hanning", fsmooth = 0.1)
}
\arguments{
\item{X}{Data frame with the following columns: 1) "sound.files": name of the .wav 
files, 2) "sel": number of the selections, 3) "start": start time of selections, 4) "end": 
end time of selections. The ouptut of \code{\link{manualoc}} or \code{\link{autodetec}} can
be used as the input data frame.}

\item{bp}{A numeric vector of length 2 for the lower and upper limits of a 
frequency bandpass filter (in kHz) or "frange" to indicate that values in low.freq
and high.freq columns will be used as bandpass limits. Default is c(0, 22).Lower limit of
 bandpass is not applied to fundamental frequencies.}

\item{wl}{A numeric vector of length 1 specifying the spectrogram window length. Default is 512. See 'wl.freq' for setting windows length independenlty in the frequency domain.}

\item{wl.freq}{A numeric vector of length 1 specifying the window length of the spectrogram
for measurements on the frecuency spectrum. Default is 512. Higher values would provide 
more accurate measurements. Note that this allows to increase measurement precision independently in the time and frequency domain. If \code{NULL} (default) then the 'wl' value is used.}

\item{threshold}{amplitude threshold (\%) for fundamental frequency and 
dominant frequency detection. Default is 15.}

\item{parallel}{Numeric. Controls whether parallel computing is applied.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{fast}{Logical. If \code{TRUE} (default) then the peakf acoustic parameter (see below) is not computed, which 
substantially increases performance (~9 times faster).}

\item{path}{Character string containing the directory path where the sound files are located. 
If \code{NULL} (default) then the current working directory is used.}

\item{pb}{Logical argument to control progress bar and messages. Default is \code{TRUE}. Note that progress bar is only used
when parallel = 1.}

\item{ovlp}{Numeric vector of length 1 specifying \% of overlap between two 
consecutive windows, used for fundamental frequency (using \code{\link[seewave]{fund}} or \code{\link[tuneR]{FF}}) and dominant frequency (using \code{\link[seewave]{dfreq}}). 
Default is 50.}

\item{ff.method}{Character. Selects the method used to calculate the fundamental
frequency. Either 'tuneR' (using \code{\link[tuneR]{FF}}) or 'seewave' (using 
\code{\link[seewave]{fund}}). Default is 'seewave'. Use trackfreqs to decide which method works the best. 
'tuneR' performs faster (and seems to be more accurate) than 'seewave'.}

\item{wn}{Character vector of length 1 specifying window name. Default is hanning'. 
See function \code{\link[seewave]{ftwindow}} for more options.}

\item{fsmooth}{A numeric vector of length 1 to smooth the frequency spectrum with a mean
sliding window (in kHz) used for mean peak frequency detection. This help to average 
amplitude "hills" to minimize the effect of amplitude modulation. Default is 0.1.}
}
\value{
Data frame with 'sound.files' and 'selec' as in the input data frame, plus the following acoustic parameters: 
\itemize{
   \item \code{duration}: length of signal (in s)
   \item \code{meanfreq}: mean frequency. Weighted average of frequency by amplitude (in kHz)
   \item \code{sd}: standard deviation of frequency weighted by amplitude
   \item \code{freq.median}: median frequency. The frequency at which the signal is divided in two frequency
   intervals of equal energy (in kHz) 
   \item \code{freq.Q25}: first quartile frequency. The frequency at which the signal is divided in two 
   frequency intervals of 25\% and 75\% energy respectively (in kHz) 
   \item \code{freq.Q75}: third quartile frequency. The frequency at which the signal is divided in two
   frequency intervals of 75\% and 25\% energy respectively (in kHz) 
   \item \code{freq.IQR}: interquartile frequency range. Frequency range between 'freq.Q25' and 'freq.Q75' 
   (in kHz) 
   \item \code{time.median}: median time. The time at which the signal is divided in two time
   intervals of equal energy (in s) 
   \item \code{time.Q25}: first quartile time. The time at which the signal is divided in two
time intervals of 25\% and 75\% energy respectively (in s). See \code{\link[seewave]{acoustat}}
   \item \code{time.Q75}: third quartile time. The time at which the signal is divided in two
   time intervals of 75\% and 25\% energy respectively (in s). See \code{\link[seewave]{acoustat}}
   \item \code{time.IQR}: interquartile time range. Time range between 'time.Q25' and 'time.Q75' 
   (in s). See \code{\link[seewave]{acoustat}}
   \item \code{skew}: skewness. Asymmetry of the spectrum (see note in \code{\link[seewave]{specprop}} description) 
   \item \code{kurt}: kurtosis. Peakedness of the spectrum (see note in \code{\link[seewave]{specprop}} description)
   \item \code{sp.ent}: spectral entropy. Energy distribution of the frequency spectrum. Pure tone ~ 0; 
   noisy ~ 1. See \code{\link[seewave]{sh}}
   \item \code{time.ent}: time entropy. Energy distribution on the time envelope. Pure tone ~ 0; 
   noisy ~ 1. See \code{\link[seewave]{th}}
   \item \code{entropy}: spectral entropy. Product of time and spectral entropy \code{sp.ent * time.ent}. 
   See \code{\link[seewave]{H}}
   \item \code{sfm}: spectral flatness. Similar to sp.ent (Pure tone ~ 0; 
   noisy ~ 1). See \code{\link[seewave]{sfm}}
   \item \code{meanfun}: average of fundamental frequency measured across the acoustic signal 
   \item \code{minfun}: minimum fundamental frequency measured across the acoustic signal 
   \item \code{maxfun}: maximum fundamental frequency measured across the acoustic signal 
   \item \code{meandom}: average of dominant frequency measured across the acoustic signal 
   \item \code{mindom}: minimum of dominant frequency measured across the acoustic signal
   \item \code{maxdom}: maximum of dominant frequency measured across the acoustic signal 
   \item \code{dfrange}: range of dominant frequency measured across the acoustic signal 
   \item \code{modindx}: modulation index. Calculated as the cumulative absolute 
     difference between adjacent measurements of dominant frequencies divided
     by the dominant frequency range. 1 means the signals is not modulated. 
   \item \code{startdom}:  dominant frequency measurement at the start of the signal 
   \item \code{enddom}: dominant frequency measurement at the end of the signal 
   \item \code{dfslope}: slope of the change in dominant through time ((enddom-startdom)/duration). Units are kHz/s.  
   \item \code{peakf}: peak frequency. Frequency with highest energy (only generated if \code{fast = FALSE})
   \item \code{meanpeakf}: Mean peak frequency. Frequency with highest energy from the 
   mean spectrum (see \code{\link[seewave]{meanspec}}). Typically more consistent than peakf.
   \item \code{low.freq}: low frequency. Low limit of frequency range (only generated if \code{frange.detec = TRUE})
   \item \code{high.freq}: high frequency. High limit of frequency range (only generated if \code{frange.detec = TRUE})
}
}
\description{
\code{specan} measures acoustic parameters on acoustic signals for which the start and end times 
are provided.
}
\details{
The ouptut of \code{\link{manualoc}} or \code{\link{autodetec}} can be used 
 directly without any additional modification. The function measures 29 acoustic parameters (if \code{fast = TRUE}) on 
 each selection in the data frame. Most parameters are produced internally by 
 \code{\link[seewave]{specprop}}, \code{\link[seewave]{fpeaks}}, \code{\link[seewave]{fund}},
 and \code{\link[seewave]{dfreq}} from the package seewave and \code{\link[tuneR]{FF}} 
 from the package tuneR. NAs are produced for fundamental and dominant 
 frequency measures when there are no amplitude values above the threshold.
}
\examples{
\dontrun{
# First set temporary folder
setwd(tempdir())

data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "selec.table"))
writeWave(Phae.long1,"Phae.long1.wav")
writeWave(Phae.long2,"Phae.long2.wav")
writeWave(Phae.long3,"Phae.long3.wav")
writeWave(Phae.long4,"Phae.long4.wav")

a <- specan(X = selec.table, bp = c(0, 22))

# using a diferent threshold
a <- specan(X = selec.table, bp = c(0, 22), threshold = 20)
# View(a)

}
}
\author{
Marcelo Araya-Salas (\email{araya-salas@cornell.edu}) and Grace Smith Vidaurre
}

