% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/specreator.R
\name{specreator}
\alias{specreator}
\title{Spectrograms of selected signals}
\usage{
specreator(X, wl = 512, flim = c(0, 22), wn = "hanning", pal
  = reverse.gray.colors.2, ovlp = 70, inner.mar = c(5, 4, 4, 2), outer.mar =
  c(0, 0, 0, 0), picsize = 1, res = 100, cexlab = 1, title = TRUE, trel = FALSE,
  propwidth = FALSE, xl=1, osci = FALSE, gr = FALSE,  sc = FALSE, line = TRUE,
  mar = 0.05, it = "jpeg")
}
\arguments{
\item{X}{Data frame with results containing columns for sound file name (sound.files),
selection number (selec), start and end time of signal (start and end), and selection comment (sel.comment).
The ouptut of \code{\link{manualoc}} can be used as the input data frame.}

\item{wl}{A numeric vector of length 1 specifying the window length of the spectrogram, default
is 512.}

\item{flim}{A numeric vector of length 2 for the frequency limit (in kHz) of
the spectrogram, as in \code{\link[seewave]{spectro}}. Default is c(0, 22).}

\item{wn}{Character vector of length 1 specifying window name. Default is
"hanning". See function \code{\link[seewave]{ftwindow}} for more options.}

\item{pal}{A color palette function to be used to assign colors in the
plot, as in \code{\link[seewave]{spectro}}. Default is reverse.gray.colors.2.}

\item{ovlp}{Numeric vector of length 1 specifying % overlap between two
consecutive windows, as in \code{\link[seewave]{spectro}}. Default is 70.}

\item{inner.mar}{Numeric vector with 4 elements, default is c(5,4,4,2).
Specifies number of lines in inner plot margins where axis labels fall,
with form c(bottom, left, top, right). See \code{\link[graphics]{par}}.}

\item{outer.mar}{Numeric vector with 4 elements, default is c(0,0,0,0).
Specifies number of lines in outer plot margins beyond axis labels, with
form c(bottom, left, top, right). See \code{\link[graphics]{par}}.}

\item{picsize}{Numeric argument of length 1. Controls relative size of
spectrogram. Default is 1. Ignored when propwidth is \code{TRUE}.}

\item{res}{Numeric argument of length 1. Controls image resolution.
Default is 100 (faster) although 300 - 400 is recommended for publication/
presentation quality.}

\item{cexlab}{Numeric vector of length 1 specifying the relative size of axis
labels. See \code{\link[seewave]{spectro}}.}

\item{title}{Logical argument to add a title to individual spectrograms.
Default is \code{TRUE}.}

\item{trel}{Logical argument to add a time axis scale relative to the wave.
Default is \code{FALSE}.}

\item{propwidth}{Logical argument to scale the width of spectrogram
proportionally to duration of the selection. Default is \code{FALSE}.}

\item{xl}{Numeric vector of length 1. A constant by which to scale
spectrogram width if propwidth = \code{TRUE}. Default is 1.}

\item{osci}{Logical argument to add an oscillogram underneath spectrogram, as
in \code{\link[seewave]{spectro}}. Default is \code{FALSE}.}

\item{gr}{Logical argument to add grid to spectrogram. Default is \code{FALSE}.}

\item{sc}{Logical argument to add amplitude scale to spectrogram, default is
\code{FALSE}.}

\item{line}{Logical argument to add red lines at start and end times of selection. Default is \code{TRUE}.}

\item{mar}{Numeric vector of length 1. Specifies the margins adjancent to the start and end points of selections,
dealineating spectrogram limits. Default is 0.05.}

\item{it}{A character vector of length 1 giving the image type to be used. Currently only
"tiff" and "jpeg" are admitted. Default is "jpeg".}
}
\value{
Image file with the spectrograms of the signals listed in the input data frame.
}
\description{
\code{specreator} creates spectrograms of signals when given start and end time.
}
\details{
This function creates spectrograms for visualization of vocalizations.
Setting inner.mar to c(4,4.5,2,1) and outer.mar to c(4,2,2,1) works well when picsize = 2 or 3.
Title font size, inner.mar and outer.mar (from mar and oma) don't work well when osci or sc = TRUE,
this may take some optimization by the user.
}
\examples{
\dontrun{
# First create empty folder
dir.create(file.path(getwd(),"temp"))
setwd(file.path(getwd(),"temp"))

data(list = c("Arre.aura", "Phae.cuvi"))
data(manualoc.df)
writeWave(Arre.aura, "Arre.aura.wav") #save sound files
writeWave(Phae.cuvi, "Phae.cuvi.wav")

# make Arre.aura and Phae.cuvi spectrograms

specreator(manualoc.df, flim = c(0, 14), inner.mar = c(4,4.5,2,1), outer.mar = c(4,2,2,1),
          picsize = 2, res = 300, cexlab = 2, mar = 0.05)
#check this folder!!
getwd()

unlink(getwd(),recursive = TRUE)
}
}
\author{
Marcelo Araya-Salas (\url{http://marceloarayasalas.weebly.com/}) and Grace Smith Vidaurre
}
\seealso{
\code{\link{trackfreqs}} for creating spectrograms to visualize
  frequency measurements by\code{\link{specan}}, \code{\link{snrspecs}} for
  creating spectrograms to optimize noise margins used in\code{\link{sig2noise}}

Other spectrogram.creators: \code{\link{snrspecs}};
  \code{\link{trackfreqs}}
}

