% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_track.R
\name{trackfreqs}
\alias{trackfreqs}
\title{alternative name for \code{\link{freq_track}}}
\usage{
trackfreqs(
  X,
  wl = 512,
  wl.freq = 512,
  flim = c(0, 22),
  wn = "hanning",
  pal = reverse.gray.colors.2,
  ovlp = 70,
  inner.mar = c(5, 4, 4, 2),
  outer.mar = c(0, 0, 0, 0),
  picsize = 1,
  res = 100,
  cexlab = 1,
  title = TRUE,
  propwidth = FALSE,
  xl = 1,
  osci = FALSE,
  gr = FALSE,
  sc = FALSE,
  bp = c(0, 22),
  cex = c(0.6, 1),
  threshold = 15,
  threshold.time = NULL,
  threshold.freq = NULL,
  contour = "both",
  col = c("#E37222B3", "#07889BB3"),
  pch = c(21, 24),
  mar = 0.05,
  lpos = "topright",
  it = "jpeg",
  parallel = 1,
  path = NULL,
  img.suffix = NULL,
  custom.contour = NULL,
  pb = TRUE,
  type = "p",
  leglab = c("Ffreq", "Dfreq"),
  col.alpha = 0.6,
  line = TRUE,
  fast.spec = FALSE,
  ff.method = "seewave",
  frange.detec = FALSE,
  fsmooth = 0.1,
  widths = c(2, 1),
  freq.continuity = NULL,
  clip.edges = 2,
  track.harm = FALSE,
  ...
)
}
\description{
alternative name for \code{\link{freq_track}}
}
\details{
see \code{\link{freq_track}} for documentation. \code{\link{trackfreqs}} will be deprecated in future versions.
}
\keyword{internal}
