% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav_2_flac.R
\name{wav_2_flac}
\alias{wav_2_flac}
\title{alternative name for \code{\link{wav2flac}}}
\usage{
wav_2_flac(
  files = NULL,
  path = NULL,
  overwrite = FALSE,
  pb = TRUE,
  parallel = 1,
  reverse = FALSE,
  compression = 5,
  flac.path = ""
)
}
\description{
alternative name for \code{\link{wav2flac}}
}
\details{
see \code{\link{wav2flac}} for documentation. \code{\link{wav_2_flac}} will be deprecated in future versions.
}
\keyword{internal}
