% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_sound_files.R
\name{wav_info}
\alias{wav_info}
\title{alternative name for \code{\link{info_sound_files}}}
\usage{
wav_info(
  path = NULL,
  parallel = 1,
  pb = TRUE,
  skip.error = FALSE,
  file.format = "\\\\.wav$|\\\\.wac$|\\\\.mp3$|\\\\.flac$"
)
}
\description{
alternative name for \code{\link{info_sound_files}}
}
\details{
see \code{\link{info_sound_files}} for documentation. \code{\link{wav_info}} will be deprecated in future versions.
}
\keyword{internal}
