% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav_dur.R
\name{wavdur}
\alias{wavdur}
\title{alternative name for \code{\link{wav_dur}}}
\usage{
wavdur(files = NULL, path = NULL)
}
\description{
alternative name for \code{\link{wav_dur}}
}
\details{
see \code{\link{wav_dur}} for documentation. \code{\link{wavdur}} will be deprecated in future versions.
}
\keyword{internal}
