% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav_2_flac.R
\name{wavdur}
\alias{wavdur}
\title{alternative name for \code{\link{duration_sound_files}}}
\usage{
wavdur(
  files = NULL,
  path = NULL,
  skip.error = FALSE,
  file.format = "\\\\.wav$|\\\\.wac$|\\\\.mp3$|\\\\.flac$"
)
}
\description{
alternative name for \code{\link{duration_sound_files}}
}
\details{
see \code{\link{duration_sound_files}} for documentation. \code{\link{wavdur}} will be deprecated in future versions.
}
\keyword{internal}
