% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration_wavs.R
\name{wavdur}
\alias{wavdur}
\title{alternative name for \code{\link{duration_sound_files}}}
\usage{
wavdur(files = NULL, path = NULL)
}
\description{
alternative name for \code{\link{duration_sound_files}}
}
\details{
see \code{\link{duration_sound_files}} for documentation. \code{\link{wavdur}} will be deprecated in future versions.
}
\keyword{internal}
