% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xc_maps.R
\name{xcmaps}
\alias{xcmaps}
\title{alternative name for \code{\link{xc_maps}}}
\usage{
xcmaps(
  X,
  img = TRUE,
  it = "jpeg",
  res = 100,
  labels = FALSE,
  path = NULL,
  leaflet.map = FALSE,
  leaflet.cluster = FALSE
)
}
\description{
alternative name for \code{\link{xc_maps}}
}
\details{
see \code{\link{xc_maps}} for documentation. \code{\link{xcmaps}} will be deprecated in future versions.
}
\keyword{internal}
