% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xcmaps_1.2.R
\name{xcmaps}
\alias{xcmaps}
\title{Maps of Xeno Canto recordings by species}
\usage{
xcmaps(X, tiff = TRUE)
}
\arguments{
\item{X}{Data frame output from querxc().}

\item{tiff}{A logical argument specifying whether a tiff file of each species
map should be returned, default is TRUE.}
}
\value{
A map of Xeno Canto recordings per species (tiff file), or a faceted
  R plot of species map(s).
}
\description{
\code{xcmaps} creates maps to visualize the geographic spread of Xeno Canto
  recordings.
}
\examples{
\dontrun{
X <- querxc("Phaethornis anthophilus", download = FALSE)
View(X)
xcmaps(X)
xcmaps(X, tiff = FALSE)
}
}
\author{
Marcelo Araya-Salas (http://marceloarayasalas.weebly.com) and Grace Smith Vidaurre
}

