% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcorr.R
\name{xcorr}
\alias{xcorr}
\title{Spectrogram cross-correlation}
\usage{
xcorr(X, wl =512, frange= NULL, ovlp=90, dens=0.9, bp= NULL, wn='hanning', 
cor.method = "pearson", parallel = 1, path = NULL, pb = TRUE, na.rm = FALSE,
 dfrange = FALSE, cor.mat = TRUE)
}
\arguments{
\item{X}{'selection_table', 'extended_selection_table' or data frame containing columns for sound files (sound.files), 
selection number (selec), and start and end time of signal (start and end).}

\item{wl}{A numeric vector of length 1 specifying the window length of the spectrogram, default 
is 512.}

\item{frange}{A numeric vector of length 2 setting the upper and lower frequency limits (in kHz) 
in which to compare the signals. Must be provided. The \code{\link{dfts}} function can be used to determine this parameter if \code{dfrange = TRUE}. This method is more adequate for pure tone
signals. Default is \code{NULL}.}

\item{ovlp}{Numeric vector of length 1 specifying \% of overlap between two 
consecutive windows, as in \code{\link[seewave]{spectro}}. Default is 90. High values of ovlp 
slow down the function but produce more accurate results.}

\item{dens}{Numeric vector of length 1 specifying the approximate density of points in which to sample amplitude. 
See \code{\link[monitoR]{makeTemplate}}. Deafult is 0.9.}

\item{bp}{A numeric vector of length 2 for the lower and upper limits of a 
frequency bandpass filter (in kHz) in which to detect dominant frequency. 
Only applied when frange is \code{NULL}. Default is \code{NULL}.}

\item{wn}{A character vector of length 1 specifying the window name as in \code{\link[seewave]{ftwindow}}.}

\item{cor.method}{A character vector of length 1 specifying the correlation method as in \code{\link[stats]{cor}}.}

\item{parallel}{Numeric. Controls whether parallel computing is applied.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{path}{Character string containing the directory path where the sound files are located. 
If \code{NULL} (default) then the current working directory is used.}

\item{pb}{Logical argument to control progress bar. Default is \code{TRUE}. Note that progress bar is not accurate as the number of pairwise comparisons decreases on 
each iteration. The first iteration runs n-1 comparisons while the last one 
only 1 (\code{n = nrow(X)}).}

\item{na.rm}{Logical. If \code{TRUE} all NAs produced when pairwise cross-correlations failed are removed from the 
results. This means that all selections with at least 1 cross-correlation that failed are excluded.}

\item{dfrange}{Logical. If \code{TRUE} the \code{\link{dfts}} function can is used to determine the frequency range in which to compare signals.}

\item{cor.mat}{Logical. If \code{TRUE} only the correlation matrix is returned. Default is \code{TRUE}.}
}
\value{
If corr.mat is \code{TRUE} the function returns a matrix with 
the maximum (peak) correlation for each pairwise comparison. Otherwise it will return a list that includes 1) a data frame with the correlation statistic for each "sliding" step, 2) a matrix with 
the maximum correlation for each pairwise comparison, and 3) the frequency range.
}
\description{
\code{xcorr} estimates the similarity of two spectrograms by means of cross-correlation
}
\details{
This function calculates the pairwise similarity of multiple signals by means of spectrogram cross-correlation.
This method "slides" one spectrogram over the other calculating a correlation of the amplitude values at each step.
The function runs pairwise cross-correlations on several signals and returns a list including the correlation statistic
for each "sliding" step as well as the maximum (peak) correlation for each pairwise comparison. To accomplish this the margins
of the signals are expanded by half the duration of the signal both before and after the provided time coordinates. 
The correlation matrix could have NA's if some of the pairwise correlation did not work (common when sound files have been modified by band-pass filters).
This function is a modified version of the \code{\link[monitoR]{corMatch}} and \code{\link[monitoR]{makeTemplate}} 
from the awesome R package `monitoR`.
}
\examples{
{
#First set temporary working directory
# setwd(tempdir())

#load data
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4","selec.table"))
writeWave(Phae.long1, "Phae.long1.wav") #save sound files
writeWave(Phae.long2, "Phae.long2.wav")
writeWave(Phae.long3, "Phae.long3.wav")
writeWave(Phae.long4, "Phae.long4.wav")

xcor <- xcorr(X = selec.table, wl = 300, frange = c(2, 9), ovlp = 90,
dens = 1, wn = 'hanning', cor.method = "pearson")

}
}
\references{
H. Khanna, S.L.L. Gaunt & D.A. McCallum (1997). Digital spectrographic 
cross-correlation: tests of sensitivity. Bioacoustics 7(3): 209-234
}
\seealso{
\code{\link{xcorr.graph}}
}
\author{
Marcelo Araya-Salas \email{araya-salas@cornell.edu})
}
